Bind an epi object to an epiparam object

Binds any epi data class in epiparameter (epidist, vb_epidist, epiparam) or data frame to an epiparam object.

Usage

bind_epiparam(epiparam, epi_obj)

Arguments

epiparam

An <epiparam> object

epi_obj

Either an <epidist>, <vb_epidist>, <epiparam> or list of <epidist> objects. It can also be a data frame as long as the columns conform to the columns of an <epiparam> object.

Value

An <epiparam> object

Details

The epiparam class holds the library of epidemiological parameters that is stored in the epiparameter R package and can be manipulated. The bind_epiparam() function allows users to add entries to the library by binding them to the bottom of an existing epiparam object loaded in R.

The <epiparam> returned by bind_epiparam() contains the matching columns of the input objects. Therefore, if one of the input object extra columns added with are not present in the other input object these will be missing from the returned object. This also applies whether binding other <epiparam> objects or <data.frames>. When binding <epidist> objects missing data fields are given a default value before binding.

Examples

eparam <- epiparam()
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "lnorm",
  prob_distribution_params = c(meanlog = 1, sdlog = 1)
)
#> Citation cannot be created as author, year, journal or title is missing
bind_epiparam(eparam, edist)
#> Epiparam object
#> Number of distributions in library: 119
#> Number of diseases: 24
#> Number of delay distributions: 96
#> Number of offspring distributions: 10
#> Number of studies in library: 58
#> <Head of library>
#>       disease         epi_distribution prob_distribution
#> 1  Adenovirus        incubation_period             lnorm
#> 2 Chikungunya        incubation_period             lnorm
#> 3    COVID-19 hospitalisation_to_death           weibull
#> 4    COVID-19 hospitalisation_to_death             lnorm
#> 5    COVID-19        incubation_period              <NA>
#> 6    COVID-19        incubation_period           weibull
#> <113 more rows & 55 more cols not shown>