Create a distribution object from distribution name and parameters

Creates an S3 class holding the distribution and parameters from the probability distribution name, its parameters and distribution truncation and discretisation.

The class holding the distribution depends on whether it is a discretised distribution. For continuous and discrete distributions S3 classes from the {distributional} package are used, for discretised continuous distributions the an S3 class from the {distcrete} package is used.

For details on the properties of the distribution classes from each respective package see their documentation (either ?distributional or ?distcrete)

Usage

create_prob_dist(prob_dist, prob_dist_params, discretise, truncation)

Arguments

prob_dist

A character string specifying the probability distribution. This should match the R naming convention of probability distributions (e.g. lognormal is lnorm, negative binomial is nbinom, and geometric is geom).

prob_dist_params

A named vector of probability distribution parameters.

discretise

A boolean logical whether the distribution is discretised. Default is FALSE which assumes a continuous probability distribution

truncation

A numeric specifying the truncation point if the inferred distribution was truncated, NA if not or unknown.

Value

An S3 class containing the probability distribution.

Details

Truncation is enabled only for continuous distributions as there is no truncation implemented in {distcrete}.

Examples

# \donttest{
# example with continuous distribution without truncation
epiparameter:::create_prob_dist(
  prob_dist = "gamma",
  prob_dist_params = c(shape = 1, scale = 1),
  discretise = FALSE,
  truncation = NA
)
#> <distribution[1]>
#> [1] Γ(1, 1)

# example with continuous distribution with truncation
epiparameter:::create_prob_dist(
  prob_dist = "gamma",
  prob_dist_params = c(shape = 1, scale = 1),
  discretise = FALSE,
  truncation = 10
)
#> <distribution[1]>
#> [1] Γ(1, 1)[-Inf,10]

# example with discrete distribution
epiparameter:::create_prob_dist(
  prob_dist = "gamma",
  prob_dist_params = c(shape = 1, scale = 1),
  discretise = TRUE,
  truncation = NA
)
#> A discrete distribution
#>   name: gamma
#>   parameters:
#>     shape: 1
#>     scale: 1
# }