Distribution (PDF, CDF, PMF) and its parameters for epidemiological characteristics of a pathogen

Parametric probability distribution (lognormal, gamma, or weibull) for the epidemiological characteristics (incubation period, onset to admission, onset to death, and serial interval) of various pathogens. The distributions and associated best-fit parameters can be used to inform modelling and estimation of other epidemiological metrics (e.g. reproduction number) and public health decisions (e.g. quarantine period).

Usage

epidist(
  pathogen,
  delay_dist = c("incubation", "onset_to_admission", "onset_to_death", "serial_interval",
    "generation_time"),
  study = NULL
)

Arguments

pathogen

A character specifying pathogen of interest

delay_dist

A character specifying type of parameter: "incubation", "onset_to_admission", "onset_to_death"

study

A character specifying dataset to use. Defaults to study with largest sample size.

Value

Produces an object of class epidist which is composed of:

pathogen: The pathogen of interest

dist: The best-fit distribution from the literature, either from the study with the largest sample size (default) or a specific study (using the study argument)

delay_dist: The metric of interest, could be either "incubation", "onset_to_admission", or "onset_to_death"

param: The parameters of the distribution

pmf, pdf, cdf: The distributions: probability mass function (pmf), probability density function (pdf) and cumulative distribution function (cdf)

Author

Adam Kucharski

Examples

# list_distributions() will show which pathogens are available for each
# metric here we search for which incubation periods are available
list_distributions(delay_dist = "incubation")
#>             pathogen_id    type_id        study_id year size distribution
#> 1            adenovirus incubation    Lessler_etal 2009   14        lnorm
#> 2                 ebola incubation        WHO_team 2014  500        gamma
#> 3             human_CoV incubation    Lessler_etal 2009   13        lnorm
#> 4  influenza_A_seasonal incubation    Lessler_etal 2009  122        lnorm
#> 5  influenza_B_seasonal incubation    Lessler_etal 2009   76        lnorm
#> 6       influenza_H1N1p incubation      Ghani_etal 2009   16        gamma
#> 7       influenza_H1N1p incubation      Tuite_etal 2010  316        lnorm
#> 8        influenza_H5N1 incubation    Cowling_etal 2013   27      weibull
#> 9        influenza_H7N9 incubation    Cowling_etal 2013   32      weibull
#> 10              marburg incubation          Pavlin 2014   18        gamma
#> 11              measles incubation    Lessler_etal 2009   56        lnorm
#> 12             MERS_CoV incubation     Assiri_etal 2013   23        lnorm
#> 13             MERS_CoV incubation Cauchemez_et_al 2014    7        lnorm
#> 14            monkeypox incubation           Nolen 2016   16        lnorm
#> 15            monkeypox incubation  Thornhill_etal 2022   23        lnorm
#> 16        parainfluenza incubation    Lessler_etal 2009   11        lnorm
#> 17           rhinovirus incubation    Lessler_etal 2009   28        lnorm
#> 18                  RSV incubation    Lessler_etal 2009   24        lnorm
#> 19             SARS_CoV incubation   Donnelly_etal 2003   57        gamma
#> 20             SARS_CoV incubation    Lessler_etal 2009  157        lnorm
#> 21  SARS_CoV_2_wildtype incubation    McAloon_etal 2020 1269        lnorm

# example of epidist() using incubation period for ebola
epidist(pathogen = "ebola", delay_dist = "incubation")
#> Pathogen: ebola
#> Delay Distribution: incubation
#> Distribution: gamma
#> Parameters:
#>   shape: 1.55395
#>    scale: 5.856044

# when more than one study is available in the database a study can be
# specified
epidist(
  pathogen = "MERS_CoV",
  delay_dist = "incubation",
  study = "Cauchemez_et_al"
)
#> Pathogen: MERS_CoV
#> Delay Distribution: incubation
#> Distribution: lnorm
#> Parameters:
#>   meanlog: 1.6108401
#>    sdlog: 0.4333775

# example using onset to admission as the metric
epidist(pathogen = "ebola", delay_dist = "onset_to_death")
#> Pathogen: ebola
#> Delay Distribution: onset_to_death
#> Distribution: gamma
#> Parameters:
#>   shape: 2.4
#>    scale: 3.333