Create an <epidist> object(s) directly from the epiparameter library (database)

Extract an <epidist> object(s) directly from the library of epidemiological parameters. This bypasses the need to read in an <epiparam> object and convert to an <epidist> object.

If a distribution from a specific study is required, the author argument can be specified.

Multiple entries (<epidist> objects) can be returned, use the arguments to subset entries and use single_epidist = TRUE to force a single <epidist> to be returned.

Usage

epidist_db(
  disease,
  epi_dist = c("incubation_period", "onset_to_hospitalisation", "onset_to_death",
    "serial_interval", "generation_time", "offspring_distribution"),
  author = NULL,
  subset = NULL,
  single_epidist = FALSE
)

Arguments

disease

A character string specifying the disease.

epi_dist

A character string specifying the epidemiological distribution.

author

A character string specifying the author of the study reporting the distribution.

subset

Either NULL or a valid R expressions that evaluates to logicals to subset the rows of <epiparam>, or a function that can be applied directly to an <epiparam> object.

This argument allows general <data.frame> subsetting that can be combined with the subsetting done with the disease and epidist arguments (and author if specified). If left as NULL (default) no subsetting is carried out.

The expression should be specified without using the data object name (e.g. df$var) and instead just var should be supplied. In other words, this argument works the same as the subset argument in subset(). It is similar to <data-masking> using by the dplyr package.

single_epidist

A boolean logical determining whether a single <epidist> or multiple entries from the library can be returned if matched by the other arguments (disease, epi_dist, author). This argument is used to prevent multiple sets of parameters being returned when only one is wanted.

Note: If multiple entries match the arguments supplied and single_epidist = TRUE then the <epidist> that is parameterised and has the largest sample size will be returned (see is_parameterised()). If multiple entries are equal after this sorting the first entry will be returned.

Value

An <epidist> object or list of <epidist> objects.

Details

disease, epi_dist and author are given as individual arguments as these are the most common variables to subset the parameter library by. The subset argument facilitates all other subsetting of rows to select the <epidist> object(s) desired. To subset based on multiple variables separate each expression with &.

Examples

epidist_db(disease = "influenza", epi_dist = "serial_interval")
#> Using Ghani, etal (2009). “The Early Transmission Dynamics of H1N1pdm
#> Influenza in the United Kingdom.” _PLoS Currents_.
#> doi:10.1371/currents.RRN1130
#> <https://doi.org/10.1371/currents.RRN1130>. 
#> To retrieve the citation use the 'get_citation' function
#> Disease: Influenza
#> Pathogen: Influenza-A-H1N1Pdm
#> Epi Distribution: serial interval
#> Study: Ghani, etal (2009). “The Early Transmission Dynamics of H1N1pdm
#> Influenza in the United Kingdom.” _PLoS Currents_.
#> doi:10.1371/currents.RRN1130
#> <https://doi.org/10.1371/currents.RRN1130>.
#> Distribution: gamma
#> Parameters:
#>   shape: 2.622
#>   scale: 0.957

# comparison between using `epidist_db()` and `epiparam()` with
# `as_epidist()`

# load influenza serial interval from database
edist <- epidist_db(disease = "influenza", epi_dist = "serial_interval")
#> Using Ghani, etal (2009). “The Early Transmission Dynamics of H1N1pdm
#> Influenza in the United Kingdom.” _PLoS Currents_.
#> doi:10.1371/currents.RRN1130
#> <https://doi.org/10.1371/currents.RRN1130>. 
#> To retrieve the citation use the 'get_citation' function

# load database of serial intervals
eparam <- epiparam(epi_dist = "serial_interval")
# subset database to only influenza entries
eparam <- eparam[clean_disease(eparam$disease) == "influenza", ]
# convert to `epidist`
edist2 <- as_epidist(eparam)
#> Using Ghani, etal (2009). “The Early Transmission Dynamics of H1N1pdm
#> Influenza in the United Kingdom.” _PLoS Currents_.
#> doi:10.1371/currents.RRN1130
#> <https://doi.org/10.1371/currents.RRN1130>. 
#> To retrieve the citation use the 'get_citation' function
# check the two methods produce the same `epidist` object
identical(edist, edist2)
#> [1] TRUE

# example using custom subsetting
eparam <- epidist_db(
  disease = "SARS",
  epi_dist = "offspring_distribution",
  subset = sample_size > 40
)
#> Using Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>. 
#> To retrieve the citation use the 'get_citation' function

# example using functional subsetting
eparam <- epidist_db(
  disease = "COVID-19",
  epi_dist = "incubation_period",
  subset = is_parameterised
)
#> Returning 2 results that match the criteria (2 are parameterised). 
#> Use subset to filter by entry variables or single_epidist to return a single entry. 
#> To retrieve the short citation for each use the 'get_citation' function

# example forcing a single <epidist> to be returned
eparam <- epidist_db(
  disease = "SARS",
  epi_dist = "offspring_distribution",
  single_epidist = TRUE
)
#> Using list(author = list(list(given = NULL, family = "Lloyd-Smith", role = NULL, email = NULL, comment = NULL), list(given = NULL, family = "etal", role = NULL, email = NULL, comment = NULL)), year = "2005", title = "Superspreading and the effect of individual variation on disease emergence", journal = "Nature", doi = "10.1038/nature04153", pmid = "16292310"). 
#> To retrieve the short citation use the 'get_citation' function