PDF, CDF, PMF, quantiles and random number generation for <epidist> and <vb_epidist> objects

The <epidist> object holds a probability distribution which can either be a continuous or discrete distribution. These are the density, cumulative distribution, quantile and random number generation functions. These operate on any distribution that can be included in an <epidist> object.

Usage

# S3 method for class 'epidist'
density(x, at, ...)

# S3 method for class 'epidist'
cdf(x, q, ..., log = FALSE)

# S3 method for class 'epidist'
quantile(x, p, ...)

# S3 method for class 'epidist'
generate(x, times, ...)

# S3 method for class 'vb_epidist'
density(x, at, ...)

# S3 method for class 'vb_epidist'
cdf(x, q, ...)

# S3 method for class 'vb_epidist'
quantile(x, p, ...)

# S3 method for class 'vb_epidist'
generate(x, times, ...)

Arguments

x

An <epidist> or <vb_epidist> object.

at

The quantiles to evaluate at.

...

dots Extra arguments to be passed to the method.

q

The quantiles to evaluate at.

log

If TRUE, probabilities will be given as log probabilities.

p

The probabilities to evaluate at.

times

The number of random samples.

Value

If an <epidist> object is given a numeric vector is returned, if an <vb_epidist> object is given a list of two elements each with a numeric vector is returned.

Examples

edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
#> Citation cannot be created as author, year, journal or title is missing

# example of each distribution method for an `epidist` object
stats::density(edist, at = 1)
#> [1] 0.3678794
distributional::cdf(edist, q = 1)
#> [1] 0.6321206
stats::quantile(edist, p = 0.2)
#> [1] 0.2231436
distributional::generate(edist, times = 10)
#>  [1] 3.56720380 1.16790186 0.05745463 0.34040705 3.47156091 1.04366207
#>  [7] 3.02627506 0.36756952 0.09970044 0.78957870

vb_edist <- vb_epidist(
  intrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(transmission_mode = "vector_borne")
  ),
  extrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = TRUE
    )
  )
)
#> Citation cannot be created as author, year, journal or title is missing
#> Citation cannot be created as author, year, journal or title is missing

# example of each distribution method for an `vb_epidist` object
stats::density(vb_edist, at = 1)
#> $intrinsic
#> [1] 0.3678794
#> 
#> $extrinsic
#> [1] 0.3678794
#> 
distributional::cdf(vb_edist, q = 1)
#> $intrinsic
#> [1] 0.6321206
#> 
#> $extrinsic
#> [1] 0.6321206
#> 
stats::quantile(vb_edist, p = 0.2)
#> $intrinsic
#> [1] 0.2231436
#> 
#> $extrinsic
#> [1] 0.2231436
#> 
distributional::generate(vb_edist, times = 10)
#> $intrinsic
#>  [1] 0.003006878 3.549754306 0.320788501 0.057208557 1.192544692 0.977676517
#>  [7] 0.270517553 0.632583351 0.687157296 2.307000169
#> 
#> $extrinsic
#>  [1] 0.5823610 0.4097386 0.4414623 0.3029223 0.3143513 0.1501912 1.9871448
#>  [8] 0.0957137 4.3464560 0.5334439
#>