epiparameter: Library of Epidemiological Parameters with Helper Functions and Classes

Library of epidemiological parameters for infectious diseases and a set of classes and helper functions.

See also

Author

Maintainer: Joshua W. Lambert joshua.lambert@lshtm.ac.uk (ORCID) [copyright holder]

Authors:

Adam Kucharski adam.kucharski@lshtm.ac.uk (ORCID) [copyright holder]

Carmen Tamayo carmen.tamayo-cuartero@lshtm.ac.uk (ORCID)

Other contributors:

Hugo Gruson hugo.gruson@data.org (ORCID) [contributor, reviewer]

Sebastian Funk sebastian.funk@lshtm.ac.uk (ORCID) [contributor]

Pratik Gupte pratik.gupte@lshtm.ac.uk (ORCID) [reviewer]