A vector of character strings with the core column names of the epidemiological parameter data exported by the epireview R package.

A vector of character strings with the core column names of the epidemiological parameter data exported by the epireview R package.

Usage

epireview_core_cols

Format

epireview_core_cols

A character vector with 58 elements.

The data is taken as the intersection of the column names of each of the disease parameter tables in the epireview R package.

Source

https://github.com/mrc-ide/epireview