Extract citation information from <epidist> or <epiparam> objects

Extract citation information from <epidist> or <epiparam> objects

Usage

get_citation(x, ...)

Arguments

x

An <epidist> or <epiparam> object.

...

Extra arguments to be passed to the method.

Value

A single character string or list of character string citations. Length of list output is equal to number of rows in the <epiparam> object passed to the function.

Examples

# example with epidist
eparam <- epiparam()
edist <- as_epidist(eparam[12, ])
#> Using Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>. 
#> To retrieve the citation use the 'get_citation' function
get_citation(edist)
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.

# example with epiparam
eparam <- epiparam()
get_citation(eparam)
#> [[1]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-6
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-6>.
#> 
#> [[2]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[3]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[4]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[5]]
#> Alene, etal (2021). “Serial interval and incubation period of COVID-19:
#> a systematic review and meta-analysis.” _BMC Infectious Diseases_.
#> doi:10.1186/s12879-021-05950-x
#> <https://doi.org/10.1186/s12879-021-05950-x>.
#> 
#> [[6]]
#> Bui, etal (2020). “Estimation of the incubation period of COVID-19 in
#> Vietnam.” _PLoS One_. doi:10.1371/journal.pone.0243889
#> <https://doi.org/10.1371/journal.pone.0243889>.
#> 
#> [[7]]
#> Elias, etal (2021). “The incubation period of COVID-19: A
#> meta-analysis.” _International Journal of Infectious Diseases_.
#> doi:10.1016/j.ijid.2021.01.069
#> <https://doi.org/10.1016/j.ijid.2021.01.069>.
#> 
#> [[8]]
#> Lauer, etal (2020). “The Incubation Period of Coronavirus Disease 2019
#> (COVID-19) From Publicly Reported Confirmed Cases: Estimation and
#> Application.” _Annals of Internal Medicine_. doi:10.7326/M20-0504
#> <https://doi.org/10.7326/M20-0504>.
#> 
#> [[9]]
#> Lauer, etal (2020). “The Incubation Period of Coronavirus Disease 2019
#> (COVID-19) From Publicly Reported Confirmed Cases: Estimation and
#> Application.” _Annals of Internal Medicine_. doi:10.7326/M20-0504
#> <https://doi.org/10.7326/M20-0504>.
#> 
#> [[10]]
#> Lauer, etal (2020). “The Incubation Period of Coronavirus Disease 2019
#> (COVID-19) From Publicly Reported Confirmed Cases: Estimation and
#> Application.” _Annals of Internal Medicine_. doi:10.7326/M20-0504
#> <https://doi.org/10.7326/M20-0504>.
#> 
#> [[11]]
#> Lauer, etal (2020). “The Incubation Period of Coronavirus Disease 2019
#> (COVID-19) From Publicly Reported Confirmed Cases: Estimation and
#> Application.” _Annals of Internal Medicine_. doi:10.7326/M20-0504
#> <https://doi.org/10.7326/M20-0504>.
#> 
#> [[12]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[13]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[14]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[15]]
#> McAloon, etal (2020). “Incubation period of COVID-19: a rapid
#> systematic review and meta-analysis of observational research.” _BMJ
#> Open_. doi:10.1136/bmjopen-2020-039652
#> <https://doi.org/10.1136/bmjopen-2020-039652>.
#> 
#> [[16]]
#> McAloon, etal (2020). “Incubation period of COVID-19: a rapid
#> systematic review and meta-analysis of observational research.” _BMJ
#> Open_. doi:10.1136/bmjopen-2020-039652
#> <https://doi.org/10.1136/bmjopen-2020-039652>.
#> 
#> [[17]]
#> Men, etal (2020). “Estimate the incubation period of coronavirus 2019
#> (COVID-19).” _medRxiv_. doi:10.1101/2020.02.24.20027474
#> <https://doi.org/10.1101/2020.02.24.20027474>.
#> 
#> [[18]]
#> Rai, etal (2022). “Incubation period for COVID-19: a systematic review
#> and meta-analysis.” _Zeitschrift fur Gesundheitswissenschaften_.
#> doi:10.1007/s10389-021-01478-1
#> <https://doi.org/10.1007/s10389-021-01478-1>.
#> 
#> [[19]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[20]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[21]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[22]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[23]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[24]]
#> Linton, etal (2020). “Incubation Period and Other Epidemiological
#> Characteristics of 2019 Novel Coronavirus Infections with Right
#> Truncation: A Statistical Analysis of Publicly Available Case Data.”
#> _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[25]]
#> Alene, etal (2021). “Serial interval and incubation period of COVID-19:
#> a systematic review and meta-analysis.” _BMC Infectious Diseases_.
#> doi:10.1186/s12879-021-05950-x
#> <https://doi.org/10.1186/s12879-021-05950-x>.
#> 
#> [[26]]
#> Chan, Johansson (2012). “The Incubation Periods of Dengue Viruses.”
#> _PLoS One_. doi:10.1371/journal.pone.0050972
#> <https://doi.org/10.1371/journal.pone.0050972>.
#> 
#> [[27]]
#> Chan, Johansson (2012). “The Incubation Periods of Dengue Viruses.”
#> _PLoS One_. doi:10.1371/journal.pone.0050972
#> <https://doi.org/10.1371/journal.pone.0050972>.
#> 
#> [[28]]
#> Chan, Johansson (2012). “The Incubation Periods of Dengue Viruses.”
#> _PLoS One_. doi:10.1371/journal.pone.0050972
#> <https://doi.org/10.1371/journal.pone.0050972>.
#> 
#> [[29]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[30]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[31]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[32]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[33]]
#> Eichner, etal (2011). “Incubation period of ebola hemorrhagic virus
#> subtype zaire.” _Osong Public Health and Research Perspectives_.
#> doi:10.1016/j.phrp.2011.04.001
#> <https://doi.org/10.1016/j.phrp.2011.04.001>.
#> 
#> [[34]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414993
#> <https://doi.org/10.1056/NEJMc1414993>.
#> 
#> [[35]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414994
#> <https://doi.org/10.1056/NEJMc1414994>.
#> 
#> [[36]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414995
#> <https://doi.org/10.1056/NEJMc1414995>.
#> 
#> [[37]]
#> WHO, Ebola, ResponseTeam (2015). “West African Ebola Epidemic after One
#> Year — Slowing but Not Yet under Control.” _The New England Journal of
#> Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[38]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[39]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[40]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[41]]
#> Barry, etal (2018). “Outbreak of Ebola virus disease in the Democratic
#> Republic of the Congo, April–May, 2018: an epidemiological study.” _The
#> Lancet_. doi:10.1016/S0140-6736(18)31387-4
#> <https://doi.org/10.1016/S0140-6736%2818%2931387-4>.
#> 
#> [[42]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[43]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[44]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[45]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[46]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[47]]
#> WHO, Ebola, Response, Team (2015). “West African Ebola Epidemic after
#> One Year — Slowing but Not Yet under Control.” _The New England Journal
#> of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[48]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[49]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-7
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-7>.
#> 
#> [[50]]
#> Lessler, etal (2009). “Outbreak of 2009 Pandemic Influenza A (H1N1) at
#> a New York City School.” _The New England Journal of Medicine_.
#> doi:10.1056/NEJMoa0906089 <https://doi.org/10.1056/NEJMoa0906089>.
#> 
#> [[51]]
#> Ghani, etal (2009). “The Early Transmission Dynamics of H1N1pdm
#> Influenza in the United Kingdom.” _PLoS Currents_.
#> doi:10.1371/currents.RRN1130
#> <https://doi.org/10.1371/currents.RRN1130>.
#> 
#> [[52]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-9
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-9>.
#> 
#> [[53]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-10
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-10>.
#> 
#> [[54]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-11
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-11>.
#> 
#> [[55]]
#> Lessler, etal (2009). “Outbreak of 2009 Pandemic Influenza A (H1N1) at
#> a New York City School.” _The New England Journal of Medicine_.
#> doi:10.1056/NEJMoa0906089 <https://doi.org/10.1056/NEJMoa0906089>.
#> 
#> [[56]]
#> Nishiura, Inaba (2011). “Estimation of the incubation period of
#> influenza A (H1N1-2009) among imported cases: addressing censoring
#> using outbreak data at the origin of importation.” _Journal of
#> Theoretical Biology_. doi:10.1016/j.jtbi.2010.12.017
#> <https://doi.org/10.1016/j.jtbi.2010.12.017>.
#> 
#> [[57]]
#> Nishiura, Inaba (2011). “Estimation of the incubation period of
#> influenza A (H1N1-2009) among imported cases: addressing censoring
#> using outbreak data at the origin of importation.” _Journal of
#> Theoretical Biology_. doi:10.1016/j.jtbi.2010.12.017
#> <https://doi.org/10.1016/j.jtbi.2010.12.017>.
#> 
#> [[58]]
#> Reich, etal (2009). “Estimating incubation period distributions with
#> coarse data.” _Statistics in Medicine_. doi:10.1002/sim.3659
#> <https://doi.org/10.1002/sim.3659>.
#> 
#> [[59]]
#> Reich, etal (2009). “Estimating incubation period distributions with
#> coarse data.” _Statistics in Medicine_. doi:10.1002/sim.3659
#> <https://doi.org/10.1002/sim.3659>.
#> 
#> [[60]]
#> Tuite, etal (2010). “Estimated epidemiologic parameters and morbidity
#> associated with pandemic H1N1 influenza.” _Canadian Medical Association
#> Journal_. doi:10.1503/cmaj.091807
#> <https://doi.org/10.1503/cmaj.091807>.
#> 
#> [[61]]
#> Virlogeux, etal (2015). “Estimating the Distribution of the Incubation
#> Periods of Human Avian Influenza A(H7N9) Virus Infections.” _American
#> Journal of Epidemiology_. doi:10.1093/aje/kwv115
#> <https://doi.org/10.1093/aje/kwv115>.
#> 
#> [[62]]
#> Virlogeux, etal (2015). “Estimating the Distribution of the Incubation
#> Periods of Human Avian Influenza A(H7N9) Virus Infections.” _American
#> Journal of Epidemiology_. doi:10.1093/aje/kwv115
#> <https://doi.org/10.1093/aje/kwv115>.
#> 
#> [[63]]
#> Virlogeux, etal (2016). “Association between the Severity of Influenza
#> A(H7N9) Virus Infections and Length of the Incubation Period.” _PLoS
#> One_. doi:10.1371/journal.pone.0148506
#> <https://doi.org/10.1371/journal.pone.0148506>.
#> 
#> [[64]]
#> Virlogeux, etal (2016). “Association between the Severity of Influenza
#> A(H7N9) Virus Infections and Length of the Incubation Period.” _PLoS
#> One_. doi:10.1371/journal.pone.0148506
#> <https://doi.org/10.1371/journal.pone.0148506>.
#> 
#> [[65]]
#> Virlogeux, etal (2016). “Association between the Severity of Influenza
#> A(H7N9) Virus Infections and Length of the Incubation Period.” _PLoS
#> One_. doi:10.1371/journal.pone.0148506
#> <https://doi.org/10.1371/journal.pone.0148506>.
#> 
#> [[66]]
#> Ghani, etal (2009). “The Early Transmission Dynamics of H1N1pdm
#> Influenza in the United Kingdom.” _PLoS Currents_.
#> doi:10.1371/currents.RRN1130
#> <https://doi.org/10.1371/currents.RRN1130>.
#> 
#> [[67]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[68]]
#> Pavlin (2014). “Calculation of incubation period and serial interval
#> from multiple outbreaks of Marburg virus disease.” _BMC Research
#> Notes_. doi:10.1186/1756-0500-7-906
#> <https://doi.org/10.1186/1756-0500-7-906>.
#> 
#> [[69]]
#> Pavlin (2014). “Calculation of incubation period and serial interval
#> from multiple outbreaks of Marburg virus disease.” _BMC Research
#> Notes_. doi:10.1186/1756-0500-7-906
#> <https://doi.org/10.1186/1756-0500-7-906>.
#> 
#> [[70]]
#> Colebunders (2007). “Marburg hemorrhagic fever in Durba and Watsa,
#> Democratic Republic of the Congo: clinical documentation, features of
#> illness, and treatment.” _The Journal of Infectious Diseases_.
#> doi:10.1086/520543 <https://doi.org/10.1086/520543>.
#> 
#> [[71]]
#> Ajelli, Merler (2012). “Transmission Potential and Design of Adequate
#> Control Measures for Marburg Hemorrhagic Fever.” _PLoS One_.
#> doi:10.1371/journal.pone.0050948
#> <https://doi.org/10.1371/journal.pone.0050948>.
#> 
#> [[72]]
#> Pavlin (2014). “Calculation of incubation period and serial interval
#> from multiple outbreaks of Marburg virus disease.” _BMC Research
#> Notes_. doi:10.1186/1756-0500-7-906
#> <https://doi.org/10.1186/1756-0500-7-906>.
#> 
#> [[73]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-12
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-12>.
#> 
#> [[74]]
#> Assiri, etal (2013). “Hospital Outbreak of Middle East Respiratory
#> Syndrome Coronavirus.” _The New England Journal of Medicine_.
#> doi:10.1056/NEJMoa1306742 <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[75]]
#> Cowling, etal (2015). “Preliminary epidemiological assessment of
#> MERS-CoV outbreak in South Korea, May to June 2015.”
#> _Eurosurveillance_. doi:10.2807/1560-7917.es2015.20.25.21163
#> <https://doi.org/10.2807/1560-7917.es2015.20.25.21163>.
#> 
#> [[76]]
#> Assiri, etal (2013). “Hospital Outbreak of Middle East Respiratory
#> Syndrome Coronavirus.” _The New England Journal of Medicine_.
#> doi:10.1056/NEJMoa1306742 <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[77]]
#> Mizumoto, etal (2015). “Real-time characterization of risks of death
#> associated with the Middle East respiratory syndrome (MERS) in the
#> Republic of Korea, 2015.” _BMC Medicine_. doi:10.1186/s12916-015-0468-3
#> <https://doi.org/10.1186/s12916-015-0468-3>.
#> 
#> [[78]]
#> Assiri, etal (2013). “Hospital Outbreak of Middle East Respiratory
#> Syndrome Coronavirus.” _The New England Journal of Medicine_.
#> doi:10.1056/NEJMoa1306742 <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[79]]
#> Assiri, etal (2013). “Hospital Outbreak of Middle East Respiratory
#> Syndrome Coronavirus.” _The New England Journal of Medicine_.
#> doi:10.1056/NEJMoa1306742 <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[80]]
#> Assiri, etal (2013). “Hospital Outbreak of Middle East Respiratory
#> Syndrome Coronavirus.” _The New England Journal of Medicine_.
#> doi:10.1056/NEJMoa1306742 <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[81]]
#> Cowling, etal (2015). “Preliminary epidemiological assessment of
#> MERS-CoV outbreak in South Korea, May to June 2016.”
#> _Eurosurveillance_. doi:10.2807/1560-7917.es2015.20.25.21163
#> <https://doi.org/10.2807/1560-7917.es2015.20.25.21163>.
#> 
#> [[82]]
#> Charniga, etal (2022). “Estimating the incubation period of monkeypox
#> virus during the 2022 multi-national outbreak.” _medRxiv_.
#> doi:10.1101/2022.06.22.22276713
#> <https://doi.org/10.1101/2022.06.22.22276713>.
#> 
#> [[83]]
#> Guzetta, etal (2022). “Early Estimates of Monkeypox Incubation Period,
#> Generation Time, and Reproduction Number, Italy, May-June 2022.”
#> _Emerging Infectious Diseases_. doi:10.3201/eid2810.221126
#> <https://doi.org/10.3201/eid2810.221126>.
#> 
#> [[84]]
#> Madewell, etal (2022). “Serial interval and incubation period estimates
#> of monkeypox virus infection in 12 U.S. jurisdictions, May – August
#> 2022.” _medRxiv_. doi:10.1101/2022.10.26.22281516
#> <https://doi.org/10.1101/2022.10.26.22281516>.
#> 
#> [[85]]
#> Madewell, etal (2022). “Serial interval and incubation period estimates
#> of monkeypox virus infection in 12 U.S. jurisdictions, May – August
#> 2023.” _medRxiv_. doi:10.1101/2022.10.26.22281517
#> <https://doi.org/10.1101/2022.10.26.22281517>.
#> 
#> [[86]]
#> Miura, etal (2022). “Estimated incubation period for monkeypox cases
#> confirmed in the Netherlands, May 2022.” _Eurosurveillance_.
#> doi:10.2807/1560-7917.ES.2022.27.24.2200448
#> <https://doi.org/10.2807/1560-7917.ES.2022.27.24.2200448>.
#> 
#> [[87]]
#> Wang, etal (2022). “Serial intervals and incubation periods of the
#> monkeypox virus clades.” _Journal of Travel Medicine_.
#> doi:10.1093/jtm/taac107 <https://doi.org/10.1093/jtm/taac107>.
#> 
#> [[88]]
#> Wang, etal (2022). “Serial intervals and incubation periods of the
#> monkeypox virus clades.” _Journal of Travel Medicine_.
#> doi:10.1093/jtm/taac108 <https://doi.org/10.1093/jtm/taac108>.
#> 
#> [[89]]
#> Wang, etal (2022). “Serial intervals and incubation periods of the
#> monkeypox virus clades.” _Journal of Travel Medicine_.
#> doi:10.1093/jtm/taac109 <https://doi.org/10.1093/jtm/taac109>.
#> 
#> [[90]]
#> Wei, etal (2022). “Study and prediction of the 2022 global monkeypox
#> epidemic.” _Journal of Biosafety and Biosecurity_.
#> doi:10.1016/j.jobb.2022.12.001
#> <https://doi.org/10.1016/j.jobb.2022.12.001>.
#> 
#> [[91]]
#> Wei, etal (2022). “Study and prediction of the 2022 global monkeypox
#> epidemic.” _Journal of Biosafety and Biosecurity_.
#> doi:10.1016/j.jobb.2022.12.002
#> <https://doi.org/10.1016/j.jobb.2022.12.002>.
#> 
#> [[92]]
#> Wei, etal (2022). “Study and prediction of the 2022 global monkeypox
#> epidemic.” _Journal of Biosafety and Biosecurity_.
#> doi:10.1016/j.jobb.2022.12.003
#> <https://doi.org/10.1016/j.jobb.2022.12.003>.
#> 
#> [[93]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[94]]
#> Guo, etal (2022). “Estimation of the serial interval of monkeypox
#> during the early outbreak in 2022.” _Journal of Medical Virology_.
#> doi:10.1002/jmv.28248 <https://doi.org/10.1002/jmv.28248>.
#> 
#> [[95]]
#> Madewell, etal (2022). “Serial interval and incubation period estimates
#> of monkeypox virus infection in 12 U.S. jurisdictions, May – August
#> 2024.” _medRxiv_. doi:10.1101/2022.10.26.22281518
#> <https://doi.org/10.1101/2022.10.26.22281518>.
#> 
#> [[96]]
#> Madewell, etal (2022). “Serial interval and incubation period estimates
#> of monkeypox virus infection in 12 U.S. jurisdictions, May – August
#> 2025.” _medRxiv_. doi:10.1101/2022.10.26.22281519
#> <https://doi.org/10.1101/2022.10.26.22281519>.
#> 
#> [[97]]
#> Wang, etal (2022). “Serial intervals and incubation periods of the
#> monkeypox virus clades.” _Journal of Travel Medicine_.
#> doi:10.1093/jtm/taac105 <https://doi.org/10.1093/jtm/taac105>.
#> 
#> [[98]]
#> Wang, etal (2022). “Serial intervals and incubation periods of the
#> monkeypox virus clades.” _Journal of Travel Medicine_.
#> doi:10.1093/jtm/taac106 <https://doi.org/10.1093/jtm/taac106>.
#> 
#> [[99]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-13
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-13>.
#> 
#> [[100]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[101]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-15
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-15>.
#> 
#> [[102]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[103]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-14
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-14>.
#> 
#> [[104]]
#> Reich, etal (2009). “Estimating incubation period distributions with
#> coarse data.” _Statistics in Medicine_. doi:10.1002/sim.3659
#> <https://doi.org/10.1002/sim.3659>.
#> 
#> [[105]]
#> Reich, etal (2009). “Estimating incubation period distributions with
#> coarse data.” _Statistics in Medicine_. doi:10.1002/sim.3659
#> <https://doi.org/10.1002/sim.3659>.
#> 
#> [[106]]
#> Lessler, etal (2009). “Incubation periods of acute respiratory viral
#> infections: a systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-8
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-8>.
#> 
#> [[107]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[108]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[109]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[110]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[111]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[112]]
#> Lloyd-Smith, etal (2005). “Superspreading and the effect of individual
#> variation on disease emergence.” _Nature_. doi:10.1038/nature04153
#> <https://doi.org/10.1038/nature04153>.
#> 
#> [[113]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[114]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[115]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[116]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[117]]
#> Rudolph, etal (2014). “Incubation periods of mosquito-borne viral
#> infections: a systematic review.” _The American Journal of Tropical
#> Medicine and Hygiene_. doi:10.4269/ajtmh.13-0403
#> <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[118]]
#> Lessler, etal (2016). “Times to key events in Zika virus infection and
#> implications for blood donation: a systematic review.” _Bulletin of the
#> World Health Organization_. doi:10.2471/BLT.16.174540
#> <https://doi.org/10.2471/BLT.16.174540>.
#>