Get parameters from an <epidist> object

Extract the parameters of the distribution stored in an <epidist> object.

Usage

# S3 method for class 'epidist'
get_parameters(x, ...)

Arguments

x

An <epidist> object.

...

dots Not used, extra arguments supplied will cause a warning.

Value

A named vector of parameters or NA when the <epidist> object is unparameterised.

Details

The <epidist> object can be unparameterised in which it lacks a probability distribution or parameters of a probability distribution, in this case the get_parameters.epidist() method will return NA.

Examples

edist <- epidist_db(
  disease = "COVID-19",
  epi_dist = "serial interval",
  single_epidist = TRUE
)
#> Using Yang L, Dai J, Zhao J, Wang Y, Deng P, Wang J (2020). “Estimation of
#> incubation period and serial interval of COVID-19: analysis of 178
#> cases and 131 transmission chains in Hubei province, China.”
#> _Epidemiology and Infection_. doi:10.1017/S0950268820001338
#> <https://doi.org/10.1017/S0950268820001338>.. 
#> To retrieve the citation use the 'get_citation' function
get_parameters(edist)
#> mean   sd 
#>  4.6  4.4