Check if <epidist> or list of <epidist> objects contains a distribution and distribution parameters

Check if <epidist> or list of <epidist> objects contains a distribution and distribution parameters

Usage

is_parameterised(x, ...)

is_parameterized(x, ...)

Arguments

x

An <epidist> or list of <epidist> objects.

...

dots not used, extra arguments supplied will cause a warning.

Value

A single boolean logical for <epidist> or vector of logicals equal in length to the list of <epidist> objects input. If the <epidist>

object is missing either a probability distribution or parameters for the probability distribution returns FALSE, otherwise it returns TRUE.

Examples

# parameterised <epidist>
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
#> Citation cannot be created as author, year, journal or title is missing
is_parameterised(edist)
#> [1] TRUE

# unparameterised <epidist>
edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation"
)
#> Citation cannot be created as author, year, journal or title is missing
#> No adequate summary statistics available to calculate the parameters of the NA distribution
#> Unparameterised <epidist> object
is_parameterised(edist)
#> [1] FALSE

# list of <epidist>
edist <- epidist_db()
#> Returning 122 results that match the criteria (99 are parameterised). 
#> Use subset to filter by entry variables or single_epidist to return a single entry. 
#> To retrieve the short citation for each use the 'get_citation' function
is_parameterised(edist)
#>   [1]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [13]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE  TRUE FALSE  TRUE  TRUE
#>  [25]  TRUE  TRUE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [37]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [49]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [61]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#>  [73]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE  TRUE
#>  [85] FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE  TRUE  TRUE  TRUE FALSE  TRUE
#>  [97]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE
#> [109]  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [121] FALSE FALSE