List epidemiological distributions stored in an epiparam object

This function subsets an epiparam object to return only the chosen epidemiological distribution. The results are returned as a data frame to better see all of the returned distributions.

By default the resulting data frame is subset to only return the disease, epidemiological distribution, probability distribution, author of the study and the year of publication as well as the sample size of the study. If the all columns of the database are required set subset_db = FALSE.

Usage

list_distributions(
  epiparam,
  epi_dist = c("incubation_period", "onset_to_admission", "onset_to_death",
    "serial_interval", "generation_time", "offspring_distribution"),
  subset_db = TRUE
)

Arguments

epiparam

An epiparam object

epi_dist

A character defining parameter to be listed: "incubation", "onset_to_admission", "onset_to_death", or "serial_interval". "incubation_period" is the default epi_dist so if no epi_dist is specified the incubation periods will be returned.

subset_db

A boolean logical that determines whether a subset, defaults is TRUE.

Author

Adam Kucharski, Joshua W. Lambert

Examples

eparam <- epiparam()
list_distributions(epiparam = eparam, epi_dist = "incubation_period")
#>                  disease  epi_distribution prob_distribution         author
#> 1             adenovirus incubation_period         lognormal   Lessler_etal
#> 2            chikungunya incubation_period         lognormal   Rudolph_etal
#> 3                 dengue incubation_period         lognormal Chan_Johansson
#> 4                 dengue incubation_period         lognormal Chan_Johansson
#> 5                 dengue incubation_period         lognormal Chan_Johansson
#> 6                 dengue incubation_period         lognormal   Rudolph_etal
#> 7                 dengue incubation_period         lognormal   Rudolph_etal
#> 8      human coronavirus incubation_period         lognormal   Lessler_etal
#> 9              influenza incubation_period             gamma     Ghani_etal
#> 10             influenza incubation_period         lognormal   Lessler_etal
#> 11             influenza incubation_period         lognormal   Lessler_etal
#> 12             influenza incubation_period         lognormal   Lessler_etal
#> 13             influenza incubation_period         lognormal   Lessler_etal
#> 14             influenza incubation_period             gamma Nishiura_Inaba
#> 15             influenza incubation_period           weibull Nishiura_Inaba
#> 16             influenza incubation_period         lognormal     Reich_etal
#> 17             influenza incubation_period         lognormal     Reich_etal
#> 18             influenza incubation_period         lognormal     Tuite_etal
#> 19             influenza incubation_period           weibull Virlogeux_etal
#> 20             influenza incubation_period             gamma Virlogeux_etal
#> 21             influenza incubation_period           weibull Virlogeux_etal
#> 22             influenza incubation_period           weibull Virlogeux_etal
#> 23             influenza incubation_period           weibull Virlogeux_etal
#> 24 Japanese encephalitis incubation_period         lognormal   Rudolph_etal
#> 25 marburg virus disease incubation_period              <NA>         Pavlin
#> 26 marburg virus disease incubation_period              <NA>         Pavlin
#> 27               measles incubation_period         lognormal   Lessler_etal
#> 28         parainfluenza incubation_period         lognormal   Lessler_etal
#> 29            rhinovirus incubation_period         lognormal   Lessler_etal
#> 30     Rift Valley fever incubation_period         lognormal   Rudolph_etal
#> 31                   RSV incubation_period         lognormal   Lessler_etal
#> 32                   RSV incubation_period         lognormal     Reich_etal
#> 33                   RSV incubation_period         lognormal     Reich_etal
#> 34                  SARS incubation_period         lognormal   Lessler_etal
#> 35       west nile fever incubation_period         lognormal   Rudolph_etal
#> 36       west nile fever incubation_period         lognormal   Rudolph_etal
#> 37       west nile fever incubation_period         lognormal   Rudolph_etal
#> 38          yellow fever incubation_period         lognormal   Rudolph_etal
#> 39          yellow fever incubation_period         lognormal   Rudolph_etal
#> 40    zika virus disease incubation_period         lognormal   Lessler_etal
#>    year sample_size
#> 1  2009          14
#> 2  2014          21
#> 3  2012         146
#> 4  2012         146
#> 5  2012         153
#> 6  2014         169
#> 7  2014         124
#> 8  2009          13
#> 9  2009          16
#> 10 2009         151
#> 11 2009          90
#> 12 2009          78
#> 13 2009         134
#> 14 2011          72
#> 15 2011          72
#> 16 2009         151
#> 17 2009         151
#> 18 2010         316
#> 19 2015         229
#> 20 2015         229
#> 21 2016         395
#> 22 2016         173
#> 23 2016         222
#> 24 2014           6
#> 25 2014          76
#> 26 2014          18
#> 27 2009          55
#> 28 2009          11
#> 29 2009          28
#> 30 2014          23
#> 31 2009          24
#> 32 2009          24
#> 33 2009          24
#> 34 2009         157
#> 35 2014          18
#> 36 2014           8
#> 37 2014           6
#> 38 2014          91
#> 39 2014          80
#> 40 2016          25
# the default for list_distributions() without any arguments is to return the
# incubation period
list_distributions(epiparam = eparam)
#>                  disease  epi_distribution prob_distribution         author
#> 1             adenovirus incubation_period         lognormal   Lessler_etal
#> 2            chikungunya incubation_period         lognormal   Rudolph_etal
#> 3                 dengue incubation_period         lognormal Chan_Johansson
#> 4                 dengue incubation_period         lognormal Chan_Johansson
#> 5                 dengue incubation_period         lognormal Chan_Johansson
#> 6                 dengue incubation_period         lognormal   Rudolph_etal
#> 7                 dengue incubation_period         lognormal   Rudolph_etal
#> 8      human coronavirus incubation_period         lognormal   Lessler_etal
#> 9              influenza incubation_period             gamma     Ghani_etal
#> 10             influenza incubation_period         lognormal   Lessler_etal
#> 11             influenza incubation_period         lognormal   Lessler_etal
#> 12             influenza incubation_period         lognormal   Lessler_etal
#> 13             influenza incubation_period         lognormal   Lessler_etal
#> 14             influenza incubation_period             gamma Nishiura_Inaba
#> 15             influenza incubation_period           weibull Nishiura_Inaba
#> 16             influenza incubation_period         lognormal     Reich_etal
#> 17             influenza incubation_period         lognormal     Reich_etal
#> 18             influenza incubation_period         lognormal     Tuite_etal
#> 19             influenza incubation_period           weibull Virlogeux_etal
#> 20             influenza incubation_period             gamma Virlogeux_etal
#> 21             influenza incubation_period           weibull Virlogeux_etal
#> 22             influenza incubation_period           weibull Virlogeux_etal
#> 23             influenza incubation_period           weibull Virlogeux_etal
#> 24 Japanese encephalitis incubation_period         lognormal   Rudolph_etal
#> 25 marburg virus disease incubation_period              <NA>         Pavlin
#> 26 marburg virus disease incubation_period              <NA>         Pavlin
#> 27               measles incubation_period         lognormal   Lessler_etal
#> 28         parainfluenza incubation_period         lognormal   Lessler_etal
#> 29            rhinovirus incubation_period         lognormal   Lessler_etal
#> 30     Rift Valley fever incubation_period         lognormal   Rudolph_etal
#> 31                   RSV incubation_period         lognormal   Lessler_etal
#> 32                   RSV incubation_period         lognormal     Reich_etal
#> 33                   RSV incubation_period         lognormal     Reich_etal
#> 34                  SARS incubation_period         lognormal   Lessler_etal
#> 35       west nile fever incubation_period         lognormal   Rudolph_etal
#> 36       west nile fever incubation_period         lognormal   Rudolph_etal
#> 37       west nile fever incubation_period         lognormal   Rudolph_etal
#> 38          yellow fever incubation_period         lognormal   Rudolph_etal
#> 39          yellow fever incubation_period         lognormal   Rudolph_etal
#> 40    zika virus disease incubation_period         lognormal   Lessler_etal
#>    year sample_size
#> 1  2009          14
#> 2  2014          21
#> 3  2012         146
#> 4  2012         146
#> 5  2012         153
#> 6  2014         169
#> 7  2014         124
#> 8  2009          13
#> 9  2009          16
#> 10 2009         151
#> 11 2009          90
#> 12 2009          78
#> 13 2009         134
#> 14 2011          72
#> 15 2011          72
#> 16 2009         151
#> 17 2009         151
#> 18 2010         316
#> 19 2015         229
#> 20 2015         229
#> 21 2016         395
#> 22 2016         173
#> 23 2016         222
#> 24 2014           6
#> 25 2014          76
#> 26 2014          18
#> 27 2009          55
#> 28 2009          11
#> 29 2009          28
#> 30 2014          23
#> 31 2009          24
#> 32 2009          24
#> 33 2009          24
#> 34 2009         157
#> 35 2014          18
#> 36 2014           8
#> 37 2014           6
#> 38 2014          91
#> 39 2014          80
#> 40 2016          25
# this same process can be achieved when loading the library
eparam <- epiparam(epi_dist = "incubation_period")

# filtering for onset to death
list_distributions(epiparam = eparam, epi_dist = "onset_to_death")
#> [1] disease           epi_distribution  prob_distribution author           
#> [5] year              sample_size      
#> <0 rows> (or 0-length row.names)