Create a <vb_epidist> object

The <vb_epidist> class is an extension of the <epidist> class (although not a subclass of <epidist>). It is is used to store epidemiological parameters for vector-borne diseases. It has the same methods (print(), format(), plot(), generate(), cdf(), density(), quantile()) as the <epidist> class and therefore they are used identically.

Usage

vb_epidist(intrinsic_epidist, extrinsic_epidist)

Arguments

intrinsic_epidist

An <epidist> object.

extrinsic_epidist

An <epidist> object.

Value

<vb_epidist> object

Details

The <epidist> objects should contain metadata (epidist$metadata) indicating it is a vector-borne disease (epidist$metadata$transmission_mode = "vector_borne") and the extrinsic distribution should indicate in the metadata that it is the extrinsic distribution (epidist$metadata$extrinsic = TRUE). If these two aspects are not given the construction of the class will throw a warning.

See also

Examples

vb <- vb_epidist(
  intrinsic_epidist = epidist(
    disease = "dengue",
    pathogen = "dengue_virus",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = FALSE
    )
  ),
  extrinsic_epidist = epidist(
    disease = "dengue",
    pathogen = "dengue_virus",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 2, scale = 2),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = TRUE
    )
  )
)
#> Citation cannot be created as author, year, journal or title is missing
#> Citation cannot be created as author, year, journal or title is missing