Calculate the parameters of a probability distribution from a list of summary statistics

If parameters of a probability distribution are not provided (e.g. when describing the distribution in the literature) and instead summary statistics of a distribution are provided, the parameters can usually be calculated from the summary statistics.

This function can provide a convenient wrapper around convert_summary_stats_to_params() and extract_param() when it is not known which summary statistics can be used to calculate the parameters of a distribution.

Usage

calc_dist_params(prob_dist, prob_dist_params, summary_stats, sample_size)

Arguments

prob_dist

A character string specifying the probability distribution. This should match the R naming convention of probability distributions (e.g. lognormal is lnorm, negative binomial is nbinom, and geometric is geom).

prob_dist_params

A named vector of probability distribution parameters.

summary_stats

A list of summary statistics, use create_epidist_summary_stats() to create list. This list can include summary statistics about the inferred distribution such as it's mean and standard deviation, quantiles of the distribution, or information about the data used to fit the distribution such as lower and upper range. The summary statistics can also include uncertainty around metrics such as confidence interval around mean and standard deviation.

sample_size

The sample size of the data. Only needed when falling back on using the median-range extraction calculation.

Value

A named numeric vector with parameters.

Details

The hierarchy of methods is:

Conversion is prioritised if the mean and standard deviation are available as these are mostly analytical conversions (except for one of the Weibull conversions).

Next method if possible is extraction from percentiles. This method requires a lower percentile (between(0-50]) and an upper percentile (between (50-100)). If multiple percentiles in each of these ranges is provided the lowest value is used for the calculation.

The last method is the extraction using a median and range of the data.