Convert the summary statistics of a distribution to parameters

Convert the summary statistics for a range of distributions to the distribution's parameters. Most summary statistics are calculated analytically given the parameters. An exception is the Weibull distribution which uses a root finding numerical method.

Usage

convert_summary_stats_to_params(x, ...)

# S3 method for class 'character'
convert_summary_stats_to_params(
  x = c("lnorm", "gamma", "weibull", "nbinom", "geom"),
  ...
)

# S3 method for class 'epidist'
convert_summary_stats_to_params(x, ...)

Arguments

x

An R object.

...

<dynamic-dots> Numeric named summary statistics used to convert to parameter(s). An example is the mean and sd summary statistics for the lognormal (lnorm) distribution.

Value

A list of either one or two elements (depending on how many parameters the distribution has).

Details

Summary statistics should be named accordingly (case-sensitive):

mean: mean

median: median

mode: mode

variance: var

standard deviation: sd

coefficient of variation: cv

skewness: skewness

excess kurtosis: ex_kurtosis

Note: Not all combinations of summary statistics can be converted into distribution parameters. In this case the function will error stating that the parameters cannot be calculated from the given input.

The distribution names and parameter names follow the style of distributions in R, for example the lognormal distribution is lnorm, and its parameters are meanlog and sdlog.

Examples

# examples using characters
convert_summary_stats_to_params("lnorm", mean = 1, sd = 1)
#> $meanlog
#> [1] -0.3465736
#> 
#> $sdlog
#> [1] 0.8325546
#> 
convert_summary_stats_to_params("weibull", mean = 2, var = 2)
#> Numerical approximation used, results may be unreliable.
#> $shape
#> [1] 1.435521
#> 
#> $scale
#> [1] 2.202641
#> 
convert_summary_stats_to_params("geom", mean = 2)
#> $prob
#> [1] 0.3333333
#> 

# examples using <epidist>
epidist <- epidist_db(single_epidist = TRUE)
#> Using WHO Ebola Response Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake I,
#> Brennan R, Cori A, Donnelly C, Dorigatti I, Dye C, Eckmanns T, Ferguson
#> N, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W, Holmes D,
#> Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S, Mills H,
#> Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier L,
#> Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A, Van
#> Kerkhove M, Varsaneux O, Kannangarage N (2015). “West African Ebola
#> Epidemic after One Year — Slowing but Not Yet under Control.” _The New
#> England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.. 
#> To retrieve the citation use the 'get_citation' function
convert_summary_stats_to_params(epidist)
#> $shape
#> [1] 1.641602
#> 
#> $scale
#> [1] 4.995122
#> 

# example using <epidist> and specifying summary stats
epidist$summary_stats <- list()
convert_summary_stats_to_params(epidist, mean = 10, sd = 2)
#> $shape
#> [1] 25
#> 
#> $scale
#> [1] 0.4
#>