Create an epidist object directly from the epiparameter library (database)

This function can extract an epidist object directly from the library of epidemiological parameters without having to read in an epiparam object and pull out an epidist object from one of the entries. If a distribution from a specific study is required, the author argument can be specified. For now this must match the author entry in the database exactly to be recognised.

Usage

epidist_db(
  disease,
  epi_dist = c("incubation_period", "onset_to_admission", "onset_to_death",
    "serial_interval", "generation_time", "offspring_distribution"),
  author = NULL
)

Arguments

disease

A character string specifying the disease

epi_dist

A character string specifying the epidemiological distribution

author

The author of the study reporting the distribution

Value

An epidist object

Details

The current criteria for chosing a distribution when multiple entries exist in the database for a given disease and epidemiological distribution is to return the entry with the largest sample size.

Examples

epidist_db(disease = "influenza", epi_dist = "serial_interval")
#> Disease: influenza
#> Pathogen: influenza-A-H1N1pdm
#> Epi Distribution: serial interval
#> Study: Ghani et al. (2009) <10.1371/currents.RRN1130> PMID: 20029668
#> Distribution: gamma
#> Parameters:
#>   shape: 2.62231009365244
#>   rate: 1.04474505723205

# comparison between using `epidist_db()` and `epiparam()` with
# `as_epidist()`

# load influenza serial interval from database
edist <- epidist_db(disease = "influenza", epi_dist = "serial_interval")

# load database of serial intervals
eparam <- epiparam(epi_dist = "serial_interval")
# subset database to only influenza entries
eparam <- eparam[eparam$disease == "influenza", ]
# convert to `epidist`
edist2 <- as_epidist(eparam)
# check the two methods produce the same `epidist` object
identical(edist, edist2)
#> [1] TRUE