PDF, CDF, PMF, quantiles and random number generation for <epidist> and <vb_epidist> objects

The <epidist> object holds a probability distribution which can either be a continuous or discrete distribution. These are the density, cumulative distribution, quantile and random number generation functions. These operate on any distribution that can be included in an <epidist> object.

Usage

# S3 method for epidist
density(x, at, ...)

# S3 method for epidist
cdf(x, q, ...)

# S3 method for epidist
quantile(x, p, ...)

# S3 method for epidist
generate(x, times, ...)

# S3 method for vb_epidist
density(x, at, ...)

# S3 method for vb_epidist
cdf(x, q, ...)

# S3 method for vb_epidist
quantile(x, p, ...)

# S3 method for vb_epidist
generate(x, times, ...)

Arguments

x

An <epidist> or <vb_epidist> object.

at

The quantiles to evaluate at.

...

dots Extra arguments passed to or from other methods.

q

The quantiles to evaluate at.

p

The probabilities to evaluate at.

times

The number of random samples.

Value

If an <epidist> object is given a numeric vector is returned, if an <vb_epidist> object is given a list of two elements each with a numeric vector is returned.

Examples

edist <- epidist(
  disease = "ebola",
  epi_dist = "incubation_period",
  prob_distribution = "gamma",
  prob_distribution_params = c(shape = 1, scale = 1)
)
#> Citation cannot be created as author, year, journal or title is missing

# example of each distribution method for an `epidist` object
stats::density(edist, at = 1)
#> [1] 0.3678794
distributional::cdf(edist, q = 1)
#> [1] 0.6321206
stats::quantile(edist, p = 0.2)
#> [1] 0.2231436
distributional::generate(edist, times = 10)
#>  [1] 2.8017249 1.4206301 0.2061902 0.4236598 1.6792179 0.6080584 0.9680913
#>  [8] 2.6693617 0.3693921 1.4972126

vb_edist <- vb_epidist(
  intrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(transmission_mode = "vector_borne")
  ),
  extrinsic_epidist = epidist(
    disease = "dengue",
    epi_dist = "incubation_period",
    prob_distribution = "gamma",
    prob_distribution_params = c(shape = 1, scale = 1),
    metadata = create_epidist_metadata(
      transmission_mode = "vector_borne",
      extrinsic = TRUE
    )
  )
)
#> Citation cannot be created as author, year, journal or title is missing
#> Citation cannot be created as author, year, journal or title is missing

# example of each distribution method for an `vb_epidist` object
stats::density(vb_edist, at = 1)
#> $intrinsic
#> [1] 0.3678794
#> 
#> $extrinsic
#> [1] 0.3678794
#> 
distributional::cdf(vb_edist, q = 1)
#> $intrinsic
#> [1] 0.6321206
#> 
#> $extrinsic
#> [1] 0.6321206
#> 
stats::quantile(vb_edist, p = 0.2)
#> $intrinsic
#> [1] 0.2231436
#> 
#> $extrinsic
#> [1] 0.2231436
#> 
distributional::generate(vb_edist, times = 10)
#> $intrinsic
#>  [1] 0.45040414 0.13154798 1.33556630 0.07992511 1.03837459 1.29139155
#>  [7] 0.12601562 1.98195812 0.16814924 0.74182727
#> 
#> $extrinsic
#>  [1] 0.7115094 0.3451355 2.0954596 2.1712054 1.2170194 0.1950834 0.1595641
#>  [8] 1.5325814 0.7214311 0.3490961
#>