epiparameter: Classes and Helper Functions for Working with Epidemiological Parameters

Classes and helper functions for loading, extracting, converting, manipulating, plotting and aggregating epidemiological parameters for infectious diseases. Epidemiological parameters extracted from the literature are loaded from the 'epiparameterDB' R package.

See also

Author

Maintainer: Joshua W. Lambert joshua.lambert@lshtm.ac.uk (ORCID) [copyright holder]

Authors:

Adam Kucharski adam.kucharski@lshtm.ac.uk (ORCID) [copyright holder]

Carmen Tamayo carmen.tamayo-cuartero@lshtm.ac.uk (ORCID)

Other contributors:

Hugo Gruson hugo.gruson@data.org (ORCID) [contributor, reviewer]

Sebastian Funk sebastian.funk@lshtm.ac.uk (ORCID) [contributor]

Pratik Gupte pratik.gupte@lshtm.ac.uk (ORCID) [reviewer]

James M. Azam james.azam@lshtm.ac.uk (ORCID) [reviewer]

Chris Hartgerink chris@data.org (ORCID) [reviewer]

Tim Taylor tim.taylor@hiddenelephants.co.uk (ORCID) [reviewer]