Calculate the parameters of a parametric probability distribution from reported values of percentiles, or median and range

Summary data of distributions, as provided by reports and meta-analyses, can be used to extract the parameters of a chosen distribution. Currently available distributions are: lognormal, gamma and weibull.

Usage

extract_param(
  type = c("percentiles", "range"),
  values,
  distribution = c("lnorm", "gamma", "weibull"),
  percentiles,
  samples
)

Arguments

type

A character defining whether summary statistics based around percentiles (default) or range

values

A vector. If type = percentiles: c(percentile_1, percentile_2); and if type = range: c(median, min, max)

distribution

A character specifying distribution to use. Default is lnorm; also takes gamma and weibull.

percentiles

A vector with two elements specifying the percentiles defined in values if using type = "percentiles".

samples

A numeric specifying the sample size if using type = "range".

Value

A numeric vector with the parameter values of the distribution

Author

Adam Kucharski

Examples

# extract parameters of a lognormal distribution from the 75 percentiles
extract_param(
  type = "percentiles",
  values = c(6, 13),
  distribution = "lnorm",
  percentiles = c(0.125, 0.875)
)
#>         a         b 
#> 2.1783545 0.3360684 

# extract parameters of a gamma distribution from median and range
extract_param(
  type = "range",
  values = c(8, 6, 13),
  distribution = "gamma",
  samples = 20
)
#>        a        b 
#> 2.285959 4.076114