Extract citation information from <epidist> or list of <epidist> objects

Extract citation information from <epidist> or list of <epidist> objects

Usage

get_citation(x, ...)

Arguments

x

An <epidist> or list of <epidist> objects.

...

dots not used, extra arguments supplied will cause a warning.

Value

A single character string or list of character string citations. Length of list output is equal to number of elements in the <epidist>

object passed to the function.

Examples

# example with <epidist>
edist <- epidist_db(single_epidist = TRUE)
#> Using WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.. 
#> To retrieve the short citation use the 'get_citation' function
get_citation(edist)
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.

# example with list of <epidist>
edist <- epidist_db()
#> Returning 118 results that match the criteria (95 are parameterised). 
#> Use subset to filter by entry variables or single_epidist to return a single entry. 
#> To retrieve the short citation for each use the 'get_citation' function
get_citation(edist)
#> [[1]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-6
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-6>.
#> 
#> [[2]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-7
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-7>.
#> 
#> [[3]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-8
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-8>.
#> 
#> [[4]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-9
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-9>.
#> 
#> [[5]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-10
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-10>.
#> 
#> [[6]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-11
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-11>.
#> 
#> [[7]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-12
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-12>.
#> 
#> [[8]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-13
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-13>.
#> 
#> [[9]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-14
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-14>.
#> 
#> [[10]]
#> Lessler J, Reich NG, Brookmeyer R, Perl TM, Nelson KE, Cummings DAT
#> (2009). “Incubation periods of acute respiratory viral infections: a
#> systematic review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-15
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-15>.
#> 
#> [[11]]
#> Reich NG, Lessler J, Cummings DAT, Brookmeyer R (2009). “Estimating
#> incubation period distributions with coarse data.” _Statistics in
#> Medicine_. doi:10.1002/sim.3659 <https://doi.org/10.1002/sim.3659>.
#> 
#> [[12]]
#> Reich NG, Lessler J, Cummings DAT, Brookmeyer R (2009). “Estimating
#> incubation period distributions with coarse data.” _Statistics in
#> Medicine_. doi:10.1002/sim.3659 <https://doi.org/10.1002/sim.3659>.
#> 
#> [[13]]
#> Reich NG, Lessler J, Cummings DAT, Brookmeyer R (2009). “Estimating
#> incubation period distributions with coarse data.” _Statistics in
#> Medicine_. doi:10.1002/sim.3659 <https://doi.org/10.1002/sim.3659>.
#> 
#> [[14]]
#> Reich NG, Lessler J, Cummings DAT, Brookmeyer R (2009). “Estimating
#> incubation period distributions with coarse data.” _Statistics in
#> Medicine_. doi:10.1002/sim.3659 <https://doi.org/10.1002/sim.3659>.
#> 
#> [[15]]
#> Nishiura H, Inaba H (2011). “Estimation of the incubation period of
#> influenza A (H1N1-2009) among imported cases: addressing censoring
#> using outbreak data at the origin of importation.” _Journal of
#> Theoretical Biology_. doi:10.1016/j.jtbi.2010.12.017
#> <https://doi.org/10.1016/j.jtbi.2010.12.017>.
#> 
#> [[16]]
#> Nishiura H, Inaba H (2011). “Estimation of the incubation period of
#> influenza A (H1N1-2009) among imported cases: addressing censoring
#> using outbreak data at the origin of importation.” _Journal of
#> Theoretical Biology_. doi:10.1016/j.jtbi.2010.12.017
#> <https://doi.org/10.1016/j.jtbi.2010.12.017>.
#> 
#> [[17]]
#> Virlogeux V, Li M, Tsang TK, Feng L, Fang VJ, Jiang H, Wu P, Zheng J,
#> Lau EHY, Cao Y, Qin Y, Liao Q, Yu H, Cowling BJ (2015). “Estimating the
#> Distribution of the Incubation Periods of Human Avian Influenza A(H7N9)
#> Virus Infections.” _American Journal of Epidemiology_.
#> doi:10.1093/aje/kwv115 <https://doi.org/10.1093/aje/kwv115>.
#> 
#> [[18]]
#> Virlogeux V, Li M, Tsang TK, Feng L, Fang VJ, Jiang H, Wu P, Zheng J,
#> Lau EHY, Cao Y, Qin Y, Liao Q, Yu H, Cowling BJ (2015). “Estimating the
#> Distribution of the Incubation Periods of Human Avian Influenza A(H7N9)
#> Virus Infections.” _American Journal of Epidemiology_.
#> doi:10.1093/aje/kwv115 <https://doi.org/10.1093/aje/kwv115>.
#> 
#> [[19]]
#> Virlogeux V, Yang J, Fang VJ, Feng L, Tsang TK, Jiang H, Wu P, Zheng J,
#> Lau EHY, Qin Y, Peng Z, Peiris JSM, Yu H, Cowling BJ (2016).
#> “Association between the Severity of Influenza A(H7N9) Virus Infections
#> and Length of the Incubation Period.” _PLoS One_.
#> doi:10.1371/journal.pone.0148506
#> <https://doi.org/10.1371/journal.pone.0148506>.
#> 
#> [[20]]
#> Virlogeux V, Yang J, Fang VJ, Feng L, Tsang TK, Jiang H, Wu P, Zheng J,
#> Lau EHY, Qin Y, Peng Z, Peiris JSM, Yu H, Cowling BJ (2016).
#> “Association between the Severity of Influenza A(H7N9) Virus Infections
#> and Length of the Incubation Period.” _PLoS One_.
#> doi:10.1371/journal.pone.0148506
#> <https://doi.org/10.1371/journal.pone.0148506>.
#> 
#> [[21]]
#> Virlogeux V, Yang J, Fang VJ, Feng L, Tsang TK, Jiang H, Wu P, Zheng J,
#> Lau EHY, Qin Y, Peng Z, Peiris JSM, Yu H, Cowling BJ (2016).
#> “Association between the Severity of Influenza A(H7N9) Virus Infections
#> and Length of the Incubation Period.” _PLoS One_.
#> doi:10.1371/journal.pone.0148506
#> <https://doi.org/10.1371/journal.pone.0148506>.
#> 
#> [[22]]
#> Tuite AR, Greer AL, Whelan M, Winter A, Lee B, Yan P, Wu J, Moghadas S,
#> Buckeridge D, Pourbohloul B, Fisman DN (2010). “Estimated epidemiologic
#> parameters and morbidity associated with pandemic H1N1 influenza.”
#> _Canadian Medical Association Journal_. doi:10.1503/cmaj.091807
#> <https://doi.org/10.1503/cmaj.091807>.
#> 
#> [[23]]
#> Ghani A, Baguelin M, Griffin J, Flasche S, van Hoek AJ, Cauchemez S,
#> Donnelly C, Robertson C, White M, Truscott J, Fraser C, Garske T, White
#> P, Leach S, Hall I, Jenkins H, Ferguson N, Cooper B (2009). “The Early
#> Transmission Dynamics of H1N1pdm Influenza in the United Kingdom.”
#> _PLoS Currents_. doi:10.1371/currents.RRN1130
#> <https://doi.org/10.1371/currents.RRN1130>.
#> 
#> [[24]]
#> Ghani A, Baguelin M, Griffin J, Flasche S, van Hoek AJ, Cauchemez S,
#> Donnelly C, Robertson C, White M, Truscott J, Fraser C, Garske T, White
#> P, Leach S, Hall I, Jenkins H, Ferguson N, Cooper B (2009). “The Early
#> Transmission Dynamics of H1N1pdm Influenza in the United Kingdom.”
#> _PLoS Currents_. doi:10.1371/currents.RRN1130
#> <https://doi.org/10.1371/currents.RRN1130>.
#> 
#> [[25]]
#> Lessler J, Reich NG, Cummings DA,
#> New-York-City-Department-of-Health-and-Mental-Hygiene-Swine-Influenza-Investigation-Team
#> (2009). “Outbreak of 2009 Pandemic Influenza A (H1N1) at a New York
#> City School.” _The New England Journal of Medicine_.
#> doi:10.1056/NEJMoa0906089 <https://doi.org/10.1056/NEJMoa0906089>.
#> 
#> [[26]]
#> Lessler J, Reich NG, Cummings DA,
#> New-York-City-Department-of-Health-and-Mental-Hygiene-Swine-Influenza-Investigation-Team
#> (2009). “Outbreak of 2009 Pandemic Influenza A (H1N1) at a New York
#> City School.” _The New England Journal of Medicine_.
#> doi:10.1056/NEJMoa0906089 <https://doi.org/10.1056/NEJMoa0906089>.
#> 
#> [[27]]
#> Pavlin BI (2014). “Calculation of incubation period and serial interval
#> from multiple outbreaks of Marburg virus disease.” _BMC Research
#> Notes_. doi:10.1186/1756-0500-7-906
#> <https://doi.org/10.1186/1756-0500-7-906>.
#> 
#> [[28]]
#> Pavlin BI (2014). “Calculation of incubation period and serial interval
#> from multiple outbreaks of Marburg virus disease.” _BMC Research
#> Notes_. doi:10.1186/1756-0500-7-906
#> <https://doi.org/10.1186/1756-0500-7-906>.
#> 
#> [[29]]
#> Pavlin BI (2014). “Calculation of incubation period and serial interval
#> from multiple outbreaks of Marburg virus disease.” _BMC Research
#> Notes_. doi:10.1186/1756-0500-7-906
#> <https://doi.org/10.1186/1756-0500-7-906>.
#> 
#> [[30]]
#> Colebunders R, Tshomba A, Van Kerkhove MD, Bausch DG, Campbell P,
#> Libande M, Pirard P, Tshioko F, Mardel S, Mulangu S, Sleurs H, Rollin
#> PE, Muyembe-Tamfum J, Jeffs B, Borchert M, Scientific I, Group'
#> TC'W1MOI (2007). “Marburg hemorrhagic fever in Durba and Watsa,
#> Democratic Republic of the Congo: clinical documentation, features of
#> illness, and treatment.” _The Journal of Infectious Diseases_.
#> doi:10.1086/520543 <https://doi.org/10.1086/520543>.
#> 
#> [[31]]
#> Ajelli M, Merler S (2012). “Transmission Potential and Design of
#> Adequate Control Measures for Marburg Hemorrhagic Fever.” _PLoS One_.
#> doi:10.1371/journal.pone.0050948
#> <https://doi.org/10.1371/journal.pone.0050948>.
#> 
#> [[32]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[33]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[34]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[35]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[36]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[37]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[38]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[39]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[40]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[41]]
#> Lloyd-Smith JO, Schreiber SJ, Kopp PE, Getz WM (2005). “Superspreading
#> and the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> 
#> [[42]]
#> Chan M, Johansson MA (2012). “The Incubation Periods of Dengue
#> Viruses.” _PLoS One_. doi:10.1371/journal.pone.0050972
#> <https://doi.org/10.1371/journal.pone.0050972>.
#> 
#> [[43]]
#> Chan M, Johansson MA (2012). “The Incubation Periods of Dengue
#> Viruses.” _PLoS One_. doi:10.1371/journal.pone.0050972
#> <https://doi.org/10.1371/journal.pone.0050972>.
#> 
#> [[44]]
#> Chan M, Johansson MA (2012). “The Incubation Periods of Dengue
#> Viruses.” _PLoS One_. doi:10.1371/journal.pone.0050972
#> <https://doi.org/10.1371/journal.pone.0050972>.
#> 
#> [[45]]
#> Lessler J, Ott CT, Carcelen AC, Konikoff JM, Williamson J, Bi Q,
#> Kucirka LM, Cummings DAT, Reich NG, Chaisson LH (2016). “Times to key
#> events in Zika virus infection and implications for blood donation: a
#> systematic review.” _Bulletin of the World Health Organization_.
#> doi:10.2471/BLT.16.174540 <https://doi.org/10.2471/BLT.16.174540>.
#> 
#> [[46]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[47]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[48]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[49]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[50]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[51]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[52]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[53]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[54]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[55]]
#> Rudolph KE, Lessler J, Moloney RM, Kmush B, Cummings DAT (2014).
#> “Incubation periods of mosquito-borne viral infections: a systematic
#> review.” _The American Journal of Tropical Medicine and Hygiene_.
#> doi:10.4269/ajtmh.13-0403 <https://doi.org/10.4269/ajtmh.13-0403>.
#> 
#> [[56]]
#> Miura F, van Ewijk CE, Backer JA, Xiridou M, Franz E, de Coul EO,
#> Brandwagt D, van Cleef B, van Rijckevorsel G, Swaan C, van den Hof S,
#> Wallinga J (2022). “Estimated incubation period for monkeypox cases
#> confirmed in the Netherlands, May 2022.” _Eurosurveillance_.
#> doi:10.2807/1560-7917.ES.2022.27.24.2200448
#> <https://doi.org/10.2807/1560-7917.ES.2022.27.24.2200448>.
#> 
#> [[57]]
#> Charniga K, Masters NB, Slayton RB, Gosdin L, Minhaj FS, Philpott D,
#> Smith D, Gearhart S, Alvarado-Ramy F, Brown C, Waltenburg MA, Hughes
#> CM, Nakazawa Y (2022). “Estimating the incubation period of monkeypox
#> virus during the 2022 multi-national outbreak.” _medRxiv_.
#> doi:10.1101/2022.06.22.22276713
#> <https://doi.org/10.1101/2022.06.22.22276713>.
#> 
#> [[58]]
#> Guzzetta G, Mammone A, Ferraro F, Caraglia A, Rapiti A, Marziano V,
#> Poletti P, Cereda D, Vairo F, Mattei G, Maraglino F, Rezza G, Merler S
#> (2022). “Early Estimates of Monkeypox Incubation Period, Generation
#> Time, and Reproduction Number, Italy, May-June 2022.” _Emerging
#> Infectious Diseases_. doi:10.3201/eid2810.221126
#> <https://doi.org/10.3201/eid2810.221126>.
#> 
#> [[59]]
#> Madewell ZJ, Charniga K, Masters NB, Asher J, Fahrenwald L, Still W,
#> Chen J, Kipperman N, Bui D, Shea M, Saathoff-Huber L, Johnson S, Harbi
#> K, Berns AL, Perez T, Gateley E, Spicknall IH, Nakazawa Y, Gift TL,
#> 2022-Monkeypox-Outbreak-Response-Team (2022). “Serial interval and
#> incubation period estimates of monkeypox virus infection in 12 U.S.
#> jurisdictions, May – August 2022.” _medRxiv_.
#> doi:10.1101/2022.10.26.22281516
#> <https://doi.org/10.1101/2022.10.26.22281516>.
#> 
#> [[60]]
#> Madewell ZJ, Charniga K, Masters NB, Asher J, Fahrenwald L, Still W,
#> Chen J, Kipperman N, Bui D, Shea M, Saathoff-Huber L, Johnson S, Harbi
#> K, Berns AL, Perez T, Gateley E, Spicknall IH, Nakazawa Y, Gift TL,
#> 2022-Monkeypox-Outbreak-Response-Team (2022). “Serial interval and
#> incubation period estimates of monkeypox virus infection in 12 U.S.
#> jurisdictions, May – August 2023.” _medRxiv_.
#> doi:10.1101/2022.10.26.22281516
#> <https://doi.org/10.1101/2022.10.26.22281516>.
#> 
#> [[61]]
#> Madewell ZJ, Charniga K, Masters NB, Asher J, Fahrenwald L, Still W,
#> Chen J, Kipperman N, Bui D, Shea M, Saathoff-Huber L, Johnson S, Harbi
#> K, Berns AL, Perez T, Gateley E, Spicknall IH, Nakazawa Y, Gift TL,
#> 2022-Monkeypox-Outbreak-Response-Team (2022). “Serial interval and
#> incubation period estimates of monkeypox virus infection in 12 U.S.
#> jurisdictions, May – August 2024.” _medRxiv_.
#> doi:10.1101/2022.10.26.22281516
#> <https://doi.org/10.1101/2022.10.26.22281516>.
#> 
#> [[62]]
#> Madewell ZJ, Charniga K, Masters NB, Asher J, Fahrenwald L, Still W,
#> Chen J, Kipperman N, Bui D, Shea M, Saathoff-Huber L, Johnson S, Harbi
#> K, Berns AL, Perez T, Gateley E, Spicknall IH, Nakazawa Y, Gift TL,
#> 2022-Monkeypox-Outbreak-Response-Team (2022). “Serial interval and
#> incubation period estimates of monkeypox virus infection in 12 U.S.
#> jurisdictions, May – August 2025.” _medRxiv_.
#> doi:10.1101/2022.10.26.22281516
#> <https://doi.org/10.1101/2022.10.26.22281516>.
#> 
#> [[63]]
#> Eichner M, Dowell SF, Firese N (2011). “Incubation period of ebola
#> hemorrhagic virus subtype zaire.” _Osong Public Health and Research
#> Perspectives_. doi:10.1016/j.phrp.2011.04.001
#> <https://doi.org/10.1016/j.phrp.2011.04.001>.
#> 
#> [[64]]
#> The-Ebola-Outbreak-Epidemiology-Team, Barry A, Ahuka-Mundeke S, Ahmed
#> YA, Allarangar Y, Anoko J, Archer BN, Abedi AA, Bagaria J, Belizaire
#> MRD, Bhatia S, Bokenge T, Bruni E, Cori A, Dabire E, Diallo AM, Diallo
#> B, Donnelly CA, Dorigatti I, Dorji TC, Waeber AREC, Fall IS, Ferguson
#> NM, FitzJohn RG, Tengomo GLF, Formenty PBH, Forna A, Fortin A, Garske
#> T, Gaythorpe KA, Gurry C, Hamblion E, Djingarey MH, Haskew C, Hugonnet
#> SAL, Imai N, Impouma B, Kabongo G, Kalenga OI, Kibangou E, Lee TM,
#> Lukoya CO, Ly O, Makiala-Mandanda S, Mamba A, Mbala-Kingebeni P,
#> Mboussou FFR, Mlanda T, Makuma VM, Morgan O, Mulumba AM, Kakoni PM,
#> Mukadi-Bamuleka D, Muyembe J, Bathé NT, Ngamala PN, Ngom R, Ngoy G,
#> Nouvellet P, Nsio J, Ousman KB, Peron E, Polonsky JA, Ryan MJ, Touré A,
#> Towner R, Tshapenda G, Van De Weerdt R, Van Kerkhove M, Wendland A, Yao
#> NKM, Yoti Z, Yuma E, Kabamba GK, Mwati JdDL, Mbuy G, Lubula L, Mutombo
#> A, Mavila O, Lay Y, Kitenge E (2018). “Outbreak of Ebola virus disease
#> in the Democratic Republic of the Congo, April–May, 2018: an
#> epidemiological study.” _The Lancet_. doi:10.1016/S0140-6736(18)31387-4
#> <https://doi.org/10.1016/S0140-6736%2818%2931387-4>.
#> 
#> [[65]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[66]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[67]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[68]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[69]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[70]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[71]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[72]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[73]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[74]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[75]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[76]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[77]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[78]]
#> WHO-Ebola-Response-Team, Agua-Agum J, Ariyarajah A, Aylward B, Blake
#> IM, Brennan R, Cori A, Donnelly CA, Dorigatti I, Dye C, Eckmanns T,
#> Ferguson NM, Formenty P, Fraser C, Garcia E, Garske T, Hinsley W,
#> Holmes D, Hugonnet S, Iyengar S, Jombart T, Krishnan R, Meijers S,
#> Mills HL, Mohamed Y, Nedjati-Gilani G, Newton E, Nouvellet P, Pelletier
#> L, Perkins D, Riley S, Sagrado M, Schnitzler J, Schumacher D, Shah A,
#> Van Kerkhove MD, Varsaneux O, Kannangarage NW (2015). “West African
#> Ebola Epidemic after One Year — Slowing but Not Yet under Control.”
#> _The New England Journal of Medicine_. doi:10.1056/NEJMc1414992
#> <https://doi.org/10.1056/NEJMc1414992>.
#> 
#> [[79]]
#> Assiri A, McGeer A, Perl TM, Price CS, Rabeeah AAA, Cummings DA,
#> Alabdullatif ZN, Assad M, Almulhim A, Makhdoom H, Madani H, Alhakeem R,
#> Al-Tawfiq JA, Cotten M, Watson SJ, Kellam P, Zumla AI, Memish ZA,
#> KSA-MERSCOV-Investigation-Team (2013). “Hospital Outbreak of Middle
#> East Respiratory Syndrome Coronavirus.” _The New England Journal of
#> Medicine_. doi:10.1056/NEJMoa1306742
#> <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[80]]
#> Assiri A, McGeer A, Perl TM, Price CS, Rabeeah AAA, Cummings DA,
#> Alabdullatif ZN, Assad M, Almulhim A, Makhdoom H, Madani H, Alhakeem R,
#> Al-Tawfiq JA, Cotten M, Watson SJ, Kellam P, Zumla AI, Memish ZA,
#> KSA-MERSCOV-Investigation-Team (2013). “Hospital Outbreak of Middle
#> East Respiratory Syndrome Coronavirus.” _The New England Journal of
#> Medicine_. doi:10.1056/NEJMoa1306742
#> <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[81]]
#> Assiri A, McGeer A, Perl TM, Price CS, Rabeeah AAA, Cummings DA,
#> Alabdullatif ZN, Assad M, Almulhim A, Makhdoom H, Madani H, Alhakeem R,
#> Al-Tawfiq JA, Cotten M, Watson SJ, Kellam P, Zumla AI, Memish ZA,
#> KSA-MERSCOV-Investigation-Team (2013). “Hospital Outbreak of Middle
#> East Respiratory Syndrome Coronavirus.” _The New England Journal of
#> Medicine_. doi:10.1056/NEJMoa1306742
#> <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[82]]
#> Assiri A, McGeer A, Perl TM, Price CS, Rabeeah AAA, Cummings DA,
#> Alabdullatif ZN, Assad M, Almulhim A, Makhdoom H, Madani H, Alhakeem R,
#> Al-Tawfiq JA, Cotten M, Watson SJ, Kellam P, Zumla AI, Memish ZA,
#> KSA-MERSCOV-Investigation-Team (2013). “Hospital Outbreak of Middle
#> East Respiratory Syndrome Coronavirus.” _The New England Journal of
#> Medicine_. doi:10.1056/NEJMoa1306742
#> <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[83]]
#> Assiri A, McGeer A, Perl TM, Price CS, Rabeeah AAA, Cummings DA,
#> Alabdullatif ZN, Assad M, Almulhim A, Makhdoom H, Madani H, Alhakeem R,
#> Al-Tawfiq JA, Cotten M, Watson SJ, Kellam P, Zumla AI, Memish ZA,
#> KSA-MERSCOV-Investigation-Team (2013). “Hospital Outbreak of Middle
#> East Respiratory Syndrome Coronavirus.” _The New England Journal of
#> Medicine_. doi:10.1056/NEJMoa1306742
#> <https://doi.org/10.1056/NEJMoa1306742>.
#> 
#> [[84]]
#> Mizumoto K, Endo A, Chowell G, Miyamatsu Y, Saitoh M, Nishiura H
#> (2015). “Real-time characterization of risks of death associated with
#> the Middle East respiratory syndrome (MERS) in the Republic of Korea,
#> 2015.” _BMC Medicine_. doi:10.1186/s12916-015-0468-3
#> <https://doi.org/10.1186/s12916-015-0468-3>.
#> 
#> [[85]]
#> Cowling BJ, Park M, Fang VJ, Wu P, Leung GM, Wu JT (2015). “Preliminary
#> epidemiological assessment of MERS-CoV outbreak in South Korea, May to
#> June 2015.” _Eurosurveillance_.
#> doi:10.2807/1560-7917.es2015.20.25.21163
#> <https://doi.org/10.2807/1560-7917.es2015.20.25.21163>.
#> 
#> [[86]]
#> Cowling BJ, Park M, Fang VJ, Wu P, Leung GM, Wu JT (2015). “Preliminary
#> epidemiological assessment of MERS-CoV outbreak in South Korea, May to
#> June 2016.” _Eurosurveillance_.
#> doi:10.2807/1560-7917.es2015.20.25.21163
#> <https://doi.org/10.2807/1560-7917.es2015.20.25.21163>.
#> 
#> [[87]]
#> Men K, Wang X, Yihao L, Zhang G, Hu J, Gao Y, Han H (2020). “Estimate
#> the incubation period of coronavirus 2019 (COVID-19).” _medRxiv_.
#> doi:10.1101/2020.02.24.20027474
#> <https://doi.org/10.1101/2020.02.24.20027474>.
#> 
#> [[88]]
#> Rai B, Shukla A, Dwivedi LK (2022). “Incubation period for COVID-19: a
#> systematic review and meta-analysis.” _Zeitschrift fur
#> Gesundheitswissenschaften_. doi:10.1007/s10389-021-01478-1
#> <https://doi.org/10.1007/s10389-021-01478-1>.
#> 
#> [[89]]
#> Alene M, Yismaw L, Assemie MA, Ketema DB, Gietaneh W, Birhan TY (2021).
#> “Serial interval and incubation period of COVID-19: a systematic review
#> and meta-analysis.” _BMC Infectious Diseases_.
#> doi:10.1186/s12879-021-05950-x
#> <https://doi.org/10.1186/s12879-021-05950-x>.
#> 
#> [[90]]
#> Alene M, Yismaw L, Assemie MA, Ketema DB, Gietaneh W, Birhan TY (2021).
#> “Serial interval and incubation period of COVID-19: a systematic review
#> and meta-analysis.” _BMC Infectious Diseases_.
#> doi:10.1186/s12879-021-05950-x
#> <https://doi.org/10.1186/s12879-021-05950-x>.
#> 
#> [[91]]
#> Elias C, Sekri A, Leblanc P, Cucherat M, Vanhems P (2021). “The
#> incubation period of COVID-19: A meta-analysis.” _International Journal
#> of Infectious Diseases_. doi:10.1016/j.ijid.2021.01.069
#> <https://doi.org/10.1016/j.ijid.2021.01.069>.
#> 
#> [[92]]
#> Bui LV, Nguyen HT, Levine H, Nguyen HN, Nguyen T, Nguyen TP, Nguyen TT,
#> Do TTT, Pham NT, Bui MH (2020). “Estimation of the incubation period of
#> COVID-19 in Vietnam.” _PLoS One_. doi:10.1371/journal.pone.0243889
#> <https://doi.org/10.1371/journal.pone.0243889>.
#> 
#> [[93]]
#> McAloon C, Collins Á, Hunt K, Barber A, Byrne AW, Butler F, Casey M,
#> Griffin J, Lane E, McEvoy D, Wall P, Green M, O'Grady L, More SJ
#> (2020). “Incubation period of COVID-19: a rapid systematic review and
#> meta-analysis of observational research.” _BMJ Open_.
#> doi:10.1136/bmjopen-2020-039652
#> <https://doi.org/10.1136/bmjopen-2020-039652>.
#> 
#> [[94]]
#> McAloon C, Collins Á, Hunt K, Barber A, Byrne AW, Butler F, Casey M,
#> Griffin J, Lane E, McEvoy D, Wall P, Green M, O'Grady L, More SJ
#> (2020). “Incubation period of COVID-19: a rapid systematic review and
#> meta-analysis of observational research.” _BMJ Open_.
#> doi:10.1136/bmjopen-2020-039652
#> <https://doi.org/10.1136/bmjopen-2020-039652>.
#> 
#> [[95]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[96]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[97]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[98]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[99]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[100]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[101]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[102]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[103]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[104]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[105]]
#> Linton NM, Kobayashi T, Yang Y, Hayashi K, Akhmetzhanov AR, Jung S,
#> Yuan B, Kinoshita R, Nishiura H (2020). “Incubation Period and Other
#> Epidemiological Characteristics of 2019 Novel Coronavirus Infections
#> with Right Truncation: A Statistical Analysis of Publicly Available
#> Case Data.” _Journal of Clinical Medicine_. doi:10.3390/jcm9020538
#> <https://doi.org/10.3390/jcm9020538>.
#> 
#> [[106]]
#> Lauer SA, Grantz KH, Bi Q, Jones FK, Zheng Q, Meredith HR, Azman AS,
#> Reich NG, Lessler J (2020). “The Incubation Period of Coronavirus
#> Disease 2019 (COVID-19) From Publicly Reported Confirmed Cases:
#> Estimation and Application.” _Annals of Internal Medicine_.
#> doi:10.7326/M20-0504 <https://doi.org/10.7326/M20-0504>.
#> 
#> [[107]]
#> Lauer SA, Grantz KH, Bi Q, Jones FK, Zheng Q, Meredith HR, Azman AS,
#> Reich NG, Lessler J (2020). “The Incubation Period of Coronavirus
#> Disease 2019 (COVID-19) From Publicly Reported Confirmed Cases:
#> Estimation and Application.” _Annals of Internal Medicine_.
#> doi:10.7326/M20-0504 <https://doi.org/10.7326/M20-0504>.
#> 
#> [[108]]
#> Lauer SA, Grantz KH, Bi Q, Jones FK, Zheng Q, Meredith HR, Azman AS,
#> Reich NG, Lessler J (2020). “The Incubation Period of Coronavirus
#> Disease 2019 (COVID-19) From Publicly Reported Confirmed Cases:
#> Estimation and Application.” _Annals of Internal Medicine_.
#> doi:10.7326/M20-0504 <https://doi.org/10.7326/M20-0504>.
#> 
#> [[109]]
#> Lauer SA, Grantz KH, Bi Q, Jones FK, Zheng Q, Meredith HR, Azman AS,
#> Reich NG, Lessler J (2020). “The Incubation Period of Coronavirus
#> Disease 2019 (COVID-19) From Publicly Reported Confirmed Cases:
#> Estimation and Application.” _Annals of Internal Medicine_.
#> doi:10.7326/M20-0504 <https://doi.org/10.7326/M20-0504>.
#> 
#> [[110]]
#> Guo Z, Zhao S, Sun S, He D, Chong KC, Yeoh EK (2022). “Estimation of
#> the serial interval of monkeypox during the early outbreak in 2022.”
#> _Journal of Medical Virology_. doi:10.1002/jmv.28248
#> <https://doi.org/10.1002/jmv.28248>.
#> 
#> [[111]]
#> Wang S, Zhang F, Yuan Z, Xu M, Wang Z, Gao C, Guo R, Du Z (2022).
#> “Serial intervals and incubation periods of the monkeypox virus
#> clades.” _Journal of Travel Medicine_. doi:10.1093/jtm/taac105
#> <https://doi.org/10.1093/jtm/taac105>.
#> 
#> [[112]]
#> Wang S, Zhang F, Yuan Z, Xu M, Wang Z, Gao C, Guo R, Du Z (2022).
#> “Serial intervals and incubation periods of the monkeypox virus
#> clades.” _Journal of Travel Medicine_. doi:10.1093/jtm/taac105
#> <https://doi.org/10.1093/jtm/taac105>.
#> 
#> [[113]]
#> Wang S, Zhang F, Yuan Z, Xu M, Wang Z, Gao C, Guo R, Du Z (2022).
#> “Serial intervals and incubation periods of the monkeypox virus
#> clades.” _Journal of Travel Medicine_. doi:10.1093/jtm/taac105
#> <https://doi.org/10.1093/jtm/taac105>.
#> 
#> [[114]]
#> Wang S, Zhang F, Yuan Z, Xu M, Wang Z, Gao C, Guo R, Du Z (2022).
#> “Serial intervals and incubation periods of the monkeypox virus
#> clades.” _Journal of Travel Medicine_. doi:10.1093/jtm/taac105
#> <https://doi.org/10.1093/jtm/taac105>.
#> 
#> [[115]]
#> Wang S, Zhang F, Yuan Z, Xu M, Wang Z, Gao C, Guo R, Du Z (2022).
#> “Serial intervals and incubation periods of the monkeypox virus
#> clades.” _Journal of Travel Medicine_. doi:10.1093/jtm/taac105
#> <https://doi.org/10.1093/jtm/taac105>.
#> 
#> [[116]]
#> Wei F, Peng Z, Jin Z, Wang J, Xu X, Zhang X, Xu J, Ren Z, Bai Y, Wang
#> X, Lu B, Wang Z, Xu J, Huang S (2022). “Study and prediction of the
#> 2022 global monkeypox epidemic.” _Journal of Biosafety and
#> Biosecurity_. doi:10.1016/j.jobb.2022.12.001
#> <https://doi.org/10.1016/j.jobb.2022.12.001>.
#> 
#> [[117]]
#> Wei F, Peng Z, Jin Z, Wang J, Xu X, Zhang X, Xu J, Ren Z, Bai Y, Wang
#> X, Lu B, Wang Z, Xu J, Huang S (2022). “Study and prediction of the
#> 2022 global monkeypox epidemic.” _Journal of Biosafety and
#> Biosecurity_. doi:10.1016/j.jobb.2022.12.001
#> <https://doi.org/10.1016/j.jobb.2022.12.001>.
#> 
#> [[118]]
#> Wei F, Peng Z, Jin Z, Wang J, Xu X, Zhang X, Xu J, Ren Z, Bai Y, Wang
#> X, Lu B, Wang Z, Xu J, Huang S (2022). “Study and prediction of the
#> 2022 global monkeypox epidemic.” _Journal of Biosafety and
#> Biosecurity_. doi:10.1016/j.jobb.2022.12.001
#> <https://doi.org/10.1016/j.jobb.2022.12.001>.
#>