Method for extracting citation information from <epidist> or <epiparam> objects.

Method for extracting citation information from <epidist> or <epiparam> objects.

Usage

get_citation(x, ...)

Arguments

x

An epidist or epiparam object.

...

Extra arguments to be passed to the method.

Value

A single character string or list of character string citations. Length of list output is equal to number of rows in the epiparam object passed to the function.

Examples

# example with epidist
eparam <- epiparam()
edist <- as_epidist(eparam[12, ])
#> Using Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150. 
#> To retrieve the short citation use the 'get_citation' function
get_citation(edist)
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"

# example with epiparam
eparam <- epiparam()
get_citation(eparam)
#> [[1]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-6> PMID: 19393959"
#> 
#> [[2]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[3]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[4]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[5]]
#> [1] "Alene et al. (2021) <10.1186/s12879-021-05950-x> PMID: 33706702"
#> 
#> [[6]]
#> [1] "Bui et al. (2020) <10.1371/journal.pone.0243889> PMID: 33362233"
#> 
#> [[7]]
#> [1] "Elias et al. (2021) <10.1016/j.ijid.2021.01.069> PMID: 33548553"
#> 
#> [[8]]
#> [1] "Lauer et al. (2020) <10.7326/M20-0504> PMID: 32150748"
#> 
#> [[9]]
#> [1] "Lauer et al. (2020) <10.7326/M20-0504> PMID: 32150748"
#> 
#> [[10]]
#> [1] "Lauer et al. (2020) <10.7326/M20-0504> PMID: 32150748"
#> 
#> [[11]]
#> [1] "Lauer et al. (2020) <10.7326/M20-0504> PMID: 32150748"
#> 
#> [[12]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[13]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[14]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[15]]
#> [1] "McAloon et al. (2020) <10.1136/bmjopen-2020-039652> PMID: 32801208"
#> 
#> [[16]]
#> [1] "McAloon et al. (2020) <10.1136/bmjopen-2020-039652> PMID: 32801208"
#> 
#> [[17]]
#> [1] "Men et al. (2020) <10.1101/2020.02.24.20027474>"
#> 
#> [[18]]
#> [1] "Rai et al. (2022) <10.1007/s10389-021-01478-1> PMID: 33643779"
#> 
#> [[19]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[20]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[21]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[22]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[23]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[24]]
#> [1] "Linton et al. (2020) <10.3390/jcm9020538> PMID: 32079150"
#> 
#> [[25]]
#> [1] "Alene et al. (2021) <10.1186/s12879-021-05950-x> PMID: 33706702"
#> 
#> [[26]]
#> [1] "Chan & Johansson (2012) <10.1371/journal.pone.0050972> PMID: 23226436"
#> 
#> [[27]]
#> [1] "Chan & Johansson (2012) <10.1371/journal.pone.0050972> PMID: 23226436"
#> 
#> [[28]]
#> [1] "Chan & Johansson (2012) <10.1371/journal.pone.0050972> PMID: 23226436"
#> 
#> [[29]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[30]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[31]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[32]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[33]]
#> [1] "Eichner et al. (2011) <10.1016/j.phrp.2011.04.001> PMID: 24159443"
#> 
#> [[34]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[35]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[36]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[37]]
#> [1] "WHO Ebola ResponseTeam (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[38]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[39]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[40]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[41]]
#> [1] "Barry et al. (2018) <10.1016/S0140-6736(18)31387-4> PMID: 30047375"
#> 
#> [[42]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[43]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[44]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[45]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[46]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[47]]
#> [1] "WHO Ebola Response Team (2015) <10.1056/NEJMc1414992> PMID: 25539446"
#> 
#> [[48]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[49]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-7> PMID: 19393959"
#> 
#> [[50]]
#> [1] "Lessler et al. (2009) <10.1056/NEJMoa0906089> PMID: 20042754"
#> 
#> [[51]]
#> [1] "Ghani et al. (2009) <10.1371/currents.RRN1130> PMID: 20029668"
#> 
#> [[52]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-9> PMID: 19393959"
#> 
#> [[53]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-10> PMID: 19393959"
#> 
#> [[54]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-11> PMID: 19393959"
#> 
#> [[55]]
#> [1] "Lessler et al. (2009) <10.1056/NEJMoa0906089> PMID: 20042754"
#> 
#> [[56]]
#> [1] "Nishiura & Inaba (2011) <10.1016/j.jtbi.2010.12.017> PMID: 21168422"
#> 
#> [[57]]
#> [1] "Nishiura & Inaba (2011) <10.1016/j.jtbi.2010.12.017> PMID: 21168422"
#> 
#> [[58]]
#> [1] "Reich et al. (2009) <10.1002/sim.3659> PMID: 19598148"
#> 
#> [[59]]
#> [1] "Reich et al. (2009) <10.1002/sim.3659> PMID: 19598148"
#> 
#> [[60]]
#> [1] "Tuite et al. (2010) <10.1503/cmaj.091807> PMID: 19959592"
#> 
#> [[61]]
#> [1] "Virlogeux et al. (2015) <10.1093/aje/kwv115> PMID: 26409239"
#> 
#> [[62]]
#> [1] "Virlogeux et al. (2015) <10.1093/aje/kwv115> PMID: 26409239"
#> 
#> [[63]]
#> [1] "Virlogeux et al. (2016) <10.1371/journal.pone.0148506> PMID: 26885816"
#> 
#> [[64]]
#> [1] "Virlogeux et al. (2016) <10.1371/journal.pone.0148506> PMID: 26885816"
#> 
#> [[65]]
#> [1] "Virlogeux et al. (2016) <10.1371/journal.pone.0148506> PMID: 26885816"
#> 
#> [[66]]
#> [1] "Ghani et al. (2009) <10.1371/currents.RRN1130> PMID: 20029668"
#> 
#> [[67]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[68]]
#> [1] "Pavlin (2014) <10.1186/1756-0500-7-906> PMID: 25495697"
#> 
#> [[69]]
#> [1] "Pavlin (2014) <10.1186/1756-0500-7-906> PMID: 25495697"
#> 
#> [[70]]
#> [1] "Colebunders (2007) <10.1086/520543> PMID: 17940943"
#> 
#> [[71]]
#> [1] "Ajelli & Merler (2012) <10.1371/journal.pone.0050948> PMID: 23251407"
#> 
#> [[72]]
#> [1] "Pavlin (2014) <10.1186/1756-0500-7-906> PMID: 25495697"
#> 
#> [[73]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-12> PMID: 19393959"
#> 
#> [[74]]
#> [1] "Assiri et al. (2013) <10.1056/NEJMoa1306742> PMID: 23782161"
#> 
#> [[75]]
#> [1] "Cowling et al. (2015) <10.2807/1560-7917.es2015.20.25.21163> PMID: 26132767"
#> 
#> [[76]]
#> [1] "Assiri et al. (2013) <10.1056/NEJMoa1306742> PMID: 23782161"
#> 
#> [[77]]
#> [1] "Mizumoto et al. (2015) <10.1186/s12916-015-0468-3> PMID: 26420593"
#> 
#> [[78]]
#> [1] "Assiri et al. (2013) <10.1056/NEJMoa1306742> PMID: 23782161"
#> 
#> [[79]]
#> [1] "Assiri et al. (2013) <10.1056/NEJMoa1306742> PMID: 23782161"
#> 
#> [[80]]
#> [1] "Assiri et al. (2013) <10.1056/NEJMoa1306742> PMID: 23782161"
#> 
#> [[81]]
#> [1] "Cowling et al. (2015) <10.2807/1560-7917.es2015.20.25.21163> PMID: 26132767"
#> 
#> [[82]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[83]]
#> [1] "Charniga et al. (2022) <10.1101/2022.06.22.22276713>"
#> 
#> [[84]]
#> [1] "Guzetta et al. (2022) <10.3201/eid2810.221126> PMID: 35994726"
#> 
#> [[85]]
#> [1] "Madewell et al. (2022) <10.1101/2022.10.26.22281516>"
#> 
#> [[86]]
#> [1] "Miura et al. (2022) <10.2807/1560-7917.ES.2022.27.24.2200448> PMID: 35713026"
#> 
#> [[87]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-13> PMID: 19393959"
#> 
#> [[88]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[89]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-15> PMID: 19393959"
#> 
#> [[90]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[91]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-14> PMID: 19393959"
#> 
#> [[92]]
#> [1] "Reich et al. (2009) <10.1002/sim.3659> PMID: 19598148"
#> 
#> [[93]]
#> [1] "Reich et al. (2009) <10.1002/sim.3659> PMID: 19598148"
#> 
#> [[94]]
#> [1] "Lessler et al. (2009) <10.1016/S1473-3099(09)70069-8> PMID: 19393959"
#> 
#> [[95]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[96]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[97]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[98]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[99]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[100]]
#> [1] "Lloyd-Smith et al. (2005) <10.1038/nature04153> PMID: 16292310"
#> 
#> [[101]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[102]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[103]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[104]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[105]]
#> [1] "Rudolph et al. (2014) <10.4269/ajtmh.13-0403> PMID: 24639305"
#> 
#> [[106]]
#> [1] "Lessler et al. (2016) <10.2471/BLT.16.174540> PMID: 27821887"
#>