List epidemiological distributions stored in an epiparam object

This function subsets an epiparam object to return only the chosen epidemiological distribution. The results are returned as a data frame to better see all of the returned distributions.

By default the resulting data frame is subset to only return the disease, epidemiological distribution, probability distribution, author of the study and the year of publication as well as the sample size of the study. If the all columns of the database are required set subset_db = FALSE.

Usage

list_distributions(
  epiparam,
  epi_dist = c("incubation_period", "onset_to_hospitalisation", "onset_to_death",
    "serial_interval", "generation_time", "offspring_distribution"),
  subset_db = TRUE
)

Arguments

epiparam

An epiparam object

epi_dist

A character defining parameter to be listed: "incubation", "onset_to_hospitalisation", "onset_to_death", or "serial_interval". "incubation_period" is the default epi_dist so if no epi_dist is specified the incubation periods will be returned.

subset_db

A boolean logical that determines whether a subset, defaults is TRUE.

Author

Adam Kucharski, Joshua W. Lambert

Examples

eparam <- epiparam()
list_distributions(epiparam = eparam, epi_dist = "incubation_period")
#>                  disease  epi_distribution prob_distribution
#> 1             Adenovirus incubation_period             lnorm
#> 2            Chikungunya incubation_period             lnorm
#> 3               COVID-19 incubation_period              <NA>
#> 4               COVID-19 incubation_period           weibull
#> 5               COVID-19 incubation_period              <NA>
#> 6               COVID-19 incubation_period             lnorm
#> 7               COVID-19 incubation_period             lnorm
#> 8               COVID-19 incubation_period             lnorm
#> 9               COVID-19 incubation_period             lnorm
#> 10              COVID-19 incubation_period             lnorm
#> 11              COVID-19 incubation_period             lnorm
#> 12              COVID-19 incubation_period             lnorm
#> 13              COVID-19 incubation_period             lnorm
#> 14              COVID-19 incubation_period             lnorm
#> 15              COVID-19 incubation_period              <NA>
#> 16              COVID-19 incubation_period              <NA>
#> 17                Dengue incubation_period             lnorm
#> 18                Dengue incubation_period             lnorm
#> 19                Dengue incubation_period             lnorm
#> 20                Dengue incubation_period             lnorm
#> 21                Dengue incubation_period             lnorm
#> 22   Ebola Virus Disease incubation_period             lnorm
#> 23   Ebola Virus Disease incubation_period             gamma
#> 24   Ebola Virus Disease incubation_period             gamma
#> 25   Ebola Virus Disease incubation_period             gamma
#> 26   Ebola Virus Disease incubation_period             gamma
#> 27     Human Coronavirus incubation_period             lnorm
#> 28             Influenza incubation_period             gamma
#> 29             Influenza incubation_period             lnorm
#> 30             Influenza incubation_period             lnorm
#> 31             Influenza incubation_period             lnorm
#> 32             Influenza incubation_period             lnorm
#> 33             Influenza incubation_period             gamma
#> 34             Influenza incubation_period           weibull
#> 35             Influenza incubation_period             lnorm
#> 36             Influenza incubation_period             lnorm
#> 37             Influenza incubation_period             lnorm
#> 38             Influenza incubation_period           weibull
#> 39             Influenza incubation_period             gamma
#> 40             Influenza incubation_period           weibull
#> 41             Influenza incubation_period           weibull
#> 42             Influenza incubation_period           weibull
#> 43 Japanese Encephalitis incubation_period             lnorm
#> 44 Marburg Virus Disease incubation_period              <NA>
#> 45 Marburg Virus Disease incubation_period              <NA>
#> 46               Measles incubation_period             lnorm
#> 47                  MERS incubation_period             lnorm
#> 48                  MERS incubation_period             gamma
#> 49                  Mpox incubation_period             lnorm
#> 50                  Mpox incubation_period             gamma
#> 51                  Mpox incubation_period             lnorm
#> 52                  Mpox incubation_period             lnorm
#> 53         Parainfluenza incubation_period             lnorm
#> 54            Rhinovirus incubation_period             lnorm
#> 55     Rift Valley Fever incubation_period             lnorm
#> 56                   RSV incubation_period             lnorm
#> 57                   RSV incubation_period             lnorm
#> 58                   RSV incubation_period             lnorm
#> 59                  SARS incubation_period             lnorm
#> 60       West Nile Fever incubation_period             lnorm
#> 61       West Nile Fever incubation_period             lnorm
#> 62       West Nile Fever incubation_period             lnorm
#> 63          Yellow Fever incubation_period             lnorm
#> 64          Yellow Fever incubation_period             lnorm
#> 65    Zika Virus Disease incubation_period             lnorm
#>                     author year sample_size
#> 1             Lessler_etal 2009          14
#> 2             Rudolph_etal 2014          21
#> 3               Alene_etal 2021        1453
#> 4                 Bui_etal 2020          19
#> 5               Elias_etal 2021       28675
#> 6               Lauer_etal 2020         181
#> 7               Lauer_etal 2020          99
#> 8               Lauer_etal 2020         108
#> 9               Lauer_etal 2020          73
#> 10             Linton_etal 2020          52
#> 11             Linton_etal 2020         158
#> 12             Linton_etal 2020          52
#> 13            McAloon_etal 2020        1357
#> 14            McAloon_etal 2020        1269
#> 15                Men_etal 2020          59
#> 16                Rai_etal 2022        6241
#> 17          Chan_Johansson 2012         146
#> 18          Chan_Johansson 2012         146
#> 19          Chan_Johansson 2012         153
#> 20            Rudolph_etal 2014         169
#> 21            Rudolph_etal 2014         124
#> 22            Eichner_etal 2011         315
#> 23 WHO_Ebola_Response_Team 2015          49
#> 24 WHO_Ebola_Response_Team 2015         957
#> 25 WHO_Ebola_Response_Team 2015         792
#> 26  WHO_Ebola_ResponseTeam 2015        1798
#> 27            Lessler_etal 2009          13
#> 28              Ghani_etal 2009          16
#> 29            Lessler_etal 2009         151
#> 30            Lessler_etal 2009          90
#> 31            Lessler_etal 2009          78
#> 32            Lessler_etal 2009         124
#> 33          Nishiura_Inaba 2011          72
#> 34          Nishiura_Inaba 2011          72
#> 35              Reich_etal 2009         151
#> 36              Reich_etal 2009         151
#> 37              Tuite_etal 2010         316
#> 38          Virlogeux_etal 2015         229
#> 39          Virlogeux_etal 2015         229
#> 40          Virlogeux_etal 2016         395
#> 41          Virlogeux_etal 2016         173
#> 42          Virlogeux_etal 2016         222
#> 43            Rudolph_etal 2014           6
#> 44                  Pavlin 2014          76
#> 45                  Pavlin 2014          18
#> 46            Lessler_etal 2009          55
#> 47             Assiri_etal 2013          23
#> 48            Cowling_etal 2015         166
#> 49           Charniga_etal 2022          22
#> 50            Guzetta_etal 2022         255
#> 51           Madewell_etal 2022          35
#> 52              Miura_etal 2022          18
#> 53            Lessler_etal 2009          11
#> 54            Lessler_etal 2009          28
#> 55            Rudolph_etal 2014          23
#> 56            Lessler_etal 2009          24
#> 57              Reich_etal 2009          24
#> 58              Reich_etal 2009          24
#> 59            Lessler_etal 2009         157
#> 60            Rudolph_etal 2014          18
#> 61            Rudolph_etal 2014           8
#> 62            Rudolph_etal 2014           6
#> 63            Rudolph_etal 2014          91
#> 64            Rudolph_etal 2014          80
#> 65            Lessler_etal 2016          25
# the default for list_distributions() without any arguments is to return the
# incubation period
list_distributions(epiparam = eparam)
#>                  disease  epi_distribution prob_distribution
#> 1             Adenovirus incubation_period             lnorm
#> 2            Chikungunya incubation_period             lnorm
#> 3               COVID-19 incubation_period              <NA>
#> 4               COVID-19 incubation_period           weibull
#> 5               COVID-19 incubation_period              <NA>
#> 6               COVID-19 incubation_period             lnorm
#> 7               COVID-19 incubation_period             lnorm
#> 8               COVID-19 incubation_period             lnorm
#> 9               COVID-19 incubation_period             lnorm
#> 10              COVID-19 incubation_period             lnorm
#> 11              COVID-19 incubation_period             lnorm
#> 12              COVID-19 incubation_period             lnorm
#> 13              COVID-19 incubation_period             lnorm
#> 14              COVID-19 incubation_period             lnorm
#> 15              COVID-19 incubation_period              <NA>
#> 16              COVID-19 incubation_period              <NA>
#> 17                Dengue incubation_period             lnorm
#> 18                Dengue incubation_period             lnorm
#> 19                Dengue incubation_period             lnorm
#> 20                Dengue incubation_period             lnorm
#> 21                Dengue incubation_period             lnorm
#> 22   Ebola Virus Disease incubation_period             lnorm
#> 23   Ebola Virus Disease incubation_period             gamma
#> 24   Ebola Virus Disease incubation_period             gamma
#> 25   Ebola Virus Disease incubation_period             gamma
#> 26   Ebola Virus Disease incubation_period             gamma
#> 27     Human Coronavirus incubation_period             lnorm
#> 28             Influenza incubation_period             gamma
#> 29             Influenza incubation_period             lnorm
#> 30             Influenza incubation_period             lnorm
#> 31             Influenza incubation_period             lnorm
#> 32             Influenza incubation_period             lnorm
#> 33             Influenza incubation_period             gamma
#> 34             Influenza incubation_period           weibull
#> 35             Influenza incubation_period             lnorm
#> 36             Influenza incubation_period             lnorm
#> 37             Influenza incubation_period             lnorm
#> 38             Influenza incubation_period           weibull
#> 39             Influenza incubation_period             gamma
#> 40             Influenza incubation_period           weibull
#> 41             Influenza incubation_period           weibull
#> 42             Influenza incubation_period           weibull
#> 43 Japanese Encephalitis incubation_period             lnorm
#> 44 Marburg Virus Disease incubation_period              <NA>
#> 45 Marburg Virus Disease incubation_period              <NA>
#> 46               Measles incubation_period             lnorm
#> 47                  MERS incubation_period             lnorm
#> 48                  MERS incubation_period             gamma
#> 49                  Mpox incubation_period             lnorm
#> 50                  Mpox incubation_period             gamma
#> 51                  Mpox incubation_period             lnorm
#> 52                  Mpox incubation_period             lnorm
#> 53         Parainfluenza incubation_period             lnorm
#> 54            Rhinovirus incubation_period             lnorm
#> 55     Rift Valley Fever incubation_period             lnorm
#> 56                   RSV incubation_period             lnorm
#> 57                   RSV incubation_period             lnorm
#> 58                   RSV incubation_period             lnorm
#> 59                  SARS incubation_period             lnorm
#> 60       West Nile Fever incubation_period             lnorm
#> 61       West Nile Fever incubation_period             lnorm
#> 62       West Nile Fever incubation_period             lnorm
#> 63          Yellow Fever incubation_period             lnorm
#> 64          Yellow Fever incubation_period             lnorm
#> 65    Zika Virus Disease incubation_period             lnorm
#>                     author year sample_size
#> 1             Lessler_etal 2009          14
#> 2             Rudolph_etal 2014          21
#> 3               Alene_etal 2021        1453
#> 4                 Bui_etal 2020          19
#> 5               Elias_etal 2021       28675
#> 6               Lauer_etal 2020         181
#> 7               Lauer_etal 2020          99
#> 8               Lauer_etal 2020         108
#> 9               Lauer_etal 2020          73
#> 10             Linton_etal 2020          52
#> 11             Linton_etal 2020         158
#> 12             Linton_etal 2020          52
#> 13            McAloon_etal 2020        1357
#> 14            McAloon_etal 2020        1269
#> 15                Men_etal 2020          59
#> 16                Rai_etal 2022        6241
#> 17          Chan_Johansson 2012         146
#> 18          Chan_Johansson 2012         146
#> 19          Chan_Johansson 2012         153
#> 20            Rudolph_etal 2014         169
#> 21            Rudolph_etal 2014         124
#> 22            Eichner_etal 2011         315
#> 23 WHO_Ebola_Response_Team 2015          49
#> 24 WHO_Ebola_Response_Team 2015         957
#> 25 WHO_Ebola_Response_Team 2015         792
#> 26  WHO_Ebola_ResponseTeam 2015        1798
#> 27            Lessler_etal 2009          13
#> 28              Ghani_etal 2009          16
#> 29            Lessler_etal 2009         151
#> 30            Lessler_etal 2009          90
#> 31            Lessler_etal 2009          78
#> 32            Lessler_etal 2009         124
#> 33          Nishiura_Inaba 2011          72
#> 34          Nishiura_Inaba 2011          72
#> 35              Reich_etal 2009         151
#> 36              Reich_etal 2009         151
#> 37              Tuite_etal 2010         316
#> 38          Virlogeux_etal 2015         229
#> 39          Virlogeux_etal 2015         229
#> 40          Virlogeux_etal 2016         395
#> 41          Virlogeux_etal 2016         173
#> 42          Virlogeux_etal 2016         222
#> 43            Rudolph_etal 2014           6
#> 44                  Pavlin 2014          76
#> 45                  Pavlin 2014          18
#> 46            Lessler_etal 2009          55
#> 47             Assiri_etal 2013          23
#> 48            Cowling_etal 2015         166
#> 49           Charniga_etal 2022          22
#> 50            Guzetta_etal 2022         255
#> 51           Madewell_etal 2022          35
#> 52              Miura_etal 2022          18
#> 53            Lessler_etal 2009          11
#> 54            Lessler_etal 2009          28
#> 55            Rudolph_etal 2014          23
#> 56            Lessler_etal 2009          24
#> 57              Reich_etal 2009          24
#> 58              Reich_etal 2009          24
#> 59            Lessler_etal 2009         157
#> 60            Rudolph_etal 2014          18
#> 61            Rudolph_etal 2014           8
#> 62            Rudolph_etal 2014           6
#> 63            Rudolph_etal 2014          91
#> 64            Rudolph_etal 2014          80
#> 65            Lessler_etal 2016          25
# this same process can be achieved when loading the library
eparam <- epiparam(epi_dist = "incubation_period")

# filtering for onset to death
list_distributions(epiparam = eparam, epi_dist = "onset_to_death")
#> [1] disease           epi_distribution  prob_distribution author           
#> [5] year              sample_size      
#> <0 rows> (or 0-length row.names)