List epidemiological distributions stored in an epiparam object

This function subsets an epiparam object to return only the chosen epidemiological distribution. The results are returned as a data frame to better see all of the returned distributions.

By default the resulting data frame is subset to only return the disease, epidemiological distribution, probability distribution, author of the study and the year of publication as well as the sample size of the study. If the all columns of the database are required set subset_db = FALSE.

Usage

list_distributions(
  epiparam,
  epi_dist = c("incubation_period", "onset_to_hospitalisation", "onset_to_death",
    "serial_interval", "generation_time", "offspring_distribution"),
  subset_db = TRUE
)

Arguments

epiparam

An <epiparam> object.

epi_dist

A character defining parameter to be listed: "incubation", "onset_to_hospitalisation", "onset_to_death", or "serial_interval". "incubation_period" is the default epi_dist so if no epi_dist is specified the incubation periods will be returned.

subset_db

A boolean logical that determines whether a subset, defaults is TRUE.

Details

The <data.frame> returned by list_distributions() contains some vector columns and some list columns. For example, the author column is a list as each paper can contain multiple authors.

Author

Adam Kucharski, Joshua W. Lambert

Examples

eparam <- epiparam()
list_distributions(epiparam = eparam, epi_dist = "incubation_period")
#>                  disease  epi_distribution prob_distribution       author year
#> 1             Adenovirus incubation_period             lnorm Justin L.... 2009
#> 2            Chikungunya incubation_period             lnorm Kara E R.... 2014
#> 3               COVID-19 incubation_period              <NA> Balram R.... 2022
#> 4               COVID-19 incubation_period              <NA> Christel.... 2021
#> 5               COVID-19 incubation_period             lnorm Conor Mc.... 2020
#> 6               COVID-19 incubation_period             lnorm Conor Mc.... 2020
#> 7               COVID-19 incubation_period              <NA> Ke Men, .... 2020
#> 8               COVID-19 incubation_period           weibull Long V. .... 2020
#> 9               COVID-19 incubation_period              <NA> Muluneh .... 2021
#> 10              COVID-19 incubation_period             lnorm Natalie .... 2020
#> 11              COVID-19 incubation_period             lnorm Natalie .... 2020
#> 12              COVID-19 incubation_period             lnorm Natalie .... 2020
#> 13              COVID-19 incubation_period             lnorm Stephen .... 2020
#> 14              COVID-19 incubation_period             lnorm Stephen .... 2020
#> 15              COVID-19 incubation_period             lnorm Stephen .... 2020
#> 16              COVID-19 incubation_period             lnorm Stephen .... 2020
#> 17                Dengue incubation_period             lnorm Kara E R.... 2014
#> 18                Dengue incubation_period             lnorm Kara E R.... 2014
#> 19                Dengue incubation_period             lnorm Miranda .... 2012
#> 20                Dengue incubation_period             lnorm Miranda .... 2012
#> 21                Dengue incubation_period             lnorm Miranda .... 2012
#> 22   Ebola Virus Disease incubation_period             lnorm Martin E.... 2011
#> 23   Ebola Virus Disease incubation_period             gamma WHO-Ebol.... 2015
#> 24   Ebola Virus Disease incubation_period             gamma WHO-Ebol.... 2015
#> 25   Ebola Virus Disease incubation_period             gamma WHO-Ebol.... 2015
#> 26   Ebola Virus Disease incubation_period             gamma WHO-Ebol.... 2015
#> 27     Human Coronavirus incubation_period             lnorm Justin L.... 2009
#> 28             Influenza incubation_period             lnorm Ashleigh.... 2010
#> 29             Influenza incubation_period             gamma Azra Gha.... 2009
#> 30             Influenza incubation_period             gamma Hiroshi .... 2011
#> 31             Influenza incubation_period           weibull Hiroshi .... 2011
#> 32             Influenza incubation_period             lnorm Justin L.... 2009
#> 33             Influenza incubation_period             lnorm Justin L.... 2009
#> 34             Influenza incubation_period             lnorm Justin L.... 2009
#> 35             Influenza incubation_period             lnorm Justin L.... 2009
#> 36             Influenza incubation_period             lnorm Nicholas.... 2009
#> 37             Influenza incubation_period             lnorm Nicholas.... 2009
#> 38             Influenza incubation_period           weibull Victor V.... 2016
#> 39             Influenza incubation_period           weibull Victor V.... 2016
#> 40             Influenza incubation_period           weibull Victor V.... 2016
#> 41             Influenza incubation_period           weibull Victor V.... 2015
#> 42             Influenza incubation_period             gamma Victor V.... 2015
#> 43 Japanese Encephalitis incubation_period             lnorm Kara E R.... 2014
#> 44 Marburg Virus Disease incubation_period              <NA> Boris I..... 2014
#> 45 Marburg Virus Disease incubation_period              <NA> Boris I..... 2014
#> 46               Measles incubation_period             lnorm Justin L.... 2009
#> 47                  MERS incubation_period             lnorm Abdullah.... 2013
#> 48                  MERS incubation_period             gamma B J Cowl.... 2015
#> 49                  Mpox incubation_period              <NA> Fengying.... 2022
#> 50                  Mpox incubation_period              <NA> Fengying.... 2022
#> 51                  Mpox incubation_period              <NA> Fengying.... 2022
#> 52                  Mpox incubation_period             lnorm Fuminari.... 2022
#> 53                  Mpox incubation_period             gamma Giorgio .... 2022
#> 54                  Mpox incubation_period             lnorm Kelly Ch.... 2022
#> 55                  Mpox incubation_period              <NA> Shuqi Wa.... 2022
#> 56                  Mpox incubation_period              <NA> Shuqi Wa.... 2022
#> 57                  Mpox incubation_period              <NA> Shuqi Wa.... 2022
#> 58                  Mpox incubation_period             lnorm Zachary .... 2022
#> 59                  Mpox incubation_period             lnorm Zachary .... 2022
#> 60         Parainfluenza incubation_period             lnorm Justin L.... 2009
#> 61            Rhinovirus incubation_period             lnorm Justin L.... 2009
#> 62     Rift Valley Fever incubation_period             lnorm Kara E R.... 2014
#> 63                   RSV incubation_period             lnorm Justin L.... 2009
#> 64                   RSV incubation_period             lnorm Nicholas.... 2009
#> 65                   RSV incubation_period             lnorm Nicholas.... 2009
#> 66                  SARS incubation_period             lnorm Justin L.... 2009
#> 67       West Nile Fever incubation_period             lnorm Kara E R.... 2014
#> 68       West Nile Fever incubation_period             lnorm Kara E R.... 2014
#> 69       West Nile Fever incubation_period             lnorm Kara E R.... 2014
#> 70          Yellow Fever incubation_period             lnorm Kara E R.... 2014
#> 71          Yellow Fever incubation_period             lnorm Kara E R.... 2014
#> 72    Zika Virus Disease incubation_period             lnorm Justin L.... 2016
#>    sample_size
#> 1           14
#> 2           21
#> 3         6241
#> 4        28675
#> 5         1357
#> 6         1269
#> 7           59
#> 8           19
#> 9         1453
#> 10          52
#> 11         158
#> 12          52
#> 13         181
#> 14          99
#> 15         108
#> 16          73
#> 17         169
#> 18         124
#> 19         146
#> 20         146
#> 21         153
#> 22         196
#> 23        1798
#> 24          49
#> 25         957
#> 26         792
#> 27          13
#> 28         316
#> 29          16
#> 30          72
#> 31          72
#> 32         124
#> 33         151
#> 34          90
#> 35          78
#> 36         151
#> 37         151
#> 38         395
#> 39         173
#> 40         222
#> 41         229
#> 42         229
#> 43           6
#> 44          76
#> 45          18
#> 46          55
#> 47          23
#> 48         166
#> 49          NA
#> 50          NA
#> 51          NA
#> 52          18
#> 53         255
#> 54          22
#> 55          16
#> 56          27
#> 57         114
#> 58          35
#> 59          36
#> 60          11
#> 61          28
#> 62          23
#> 63          24
#> 64          24
#> 65          24
#> 66         157
#> 67          18
#> 68           8
#> 69           6
#> 70          91
#> 71          80
#> 72          25
# the default for list_distributions() without any arguments is to return the
# incubation period
list_distributions(epiparam = eparam)
#>                  disease  epi_distribution prob_distribution       author year
#> 1             Adenovirus incubation_period             lnorm Justin L.... 2009
#> 2            Chikungunya incubation_period             lnorm Kara E R.... 2014
#> 3               COVID-19 incubation_period              <NA> Balram R.... 2022
#> 4               COVID-19 incubation_period              <NA> Christel.... 2021
#> 5               COVID-19 incubation_period             lnorm Conor Mc.... 2020
#> 6               COVID-19 incubation_period             lnorm Conor Mc.... 2020
#> 7               COVID-19 incubation_period              <NA> Ke Men, .... 2020
#> 8               COVID-19 incubation_period           weibull Long V. .... 2020
#> 9               COVID-19 incubation_period              <NA> Muluneh .... 2021
#> 10              COVID-19 incubation_period             lnorm Natalie .... 2020
#> 11              COVID-19 incubation_period             lnorm Natalie .... 2020
#> 12              COVID-19 incubation_period             lnorm Natalie .... 2020
#> 13              COVID-19 incubation_period             lnorm Stephen .... 2020
#> 14              COVID-19 incubation_period             lnorm Stephen .... 2020
#> 15              COVID-19 incubation_period             lnorm Stephen .... 2020
#> 16              COVID-19 incubation_period             lnorm Stephen .... 2020
#> 17                Dengue incubation_period             lnorm Kara E R.... 2014
#> 18                Dengue incubation_period             lnorm Kara E R.... 2014
#> 19                Dengue incubation_period             lnorm Miranda .... 2012
#> 20                Dengue incubation_period             lnorm Miranda .... 2012
#> 21                Dengue incubation_period             lnorm Miranda .... 2012
#> 22   Ebola Virus Disease incubation_period             lnorm Martin E.... 2011
#> 23   Ebola Virus Disease incubation_period             gamma WHO-Ebol.... 2015
#> 24   Ebola Virus Disease incubation_period             gamma WHO-Ebol.... 2015
#> 25   Ebola Virus Disease incubation_period             gamma WHO-Ebol.... 2015
#> 26   Ebola Virus Disease incubation_period             gamma WHO-Ebol.... 2015
#> 27     Human Coronavirus incubation_period             lnorm Justin L.... 2009
#> 28             Influenza incubation_period             lnorm Ashleigh.... 2010
#> 29             Influenza incubation_period             gamma Azra Gha.... 2009
#> 30             Influenza incubation_period             gamma Hiroshi .... 2011
#> 31             Influenza incubation_period           weibull Hiroshi .... 2011
#> 32             Influenza incubation_period             lnorm Justin L.... 2009
#> 33             Influenza incubation_period             lnorm Justin L.... 2009
#> 34             Influenza incubation_period             lnorm Justin L.... 2009
#> 35             Influenza incubation_period             lnorm Justin L.... 2009
#> 36             Influenza incubation_period             lnorm Nicholas.... 2009
#> 37             Influenza incubation_period             lnorm Nicholas.... 2009
#> 38             Influenza incubation_period           weibull Victor V.... 2016
#> 39             Influenza incubation_period           weibull Victor V.... 2016
#> 40             Influenza incubation_period           weibull Victor V.... 2016
#> 41             Influenza incubation_period           weibull Victor V.... 2015
#> 42             Influenza incubation_period             gamma Victor V.... 2015
#> 43 Japanese Encephalitis incubation_period             lnorm Kara E R.... 2014
#> 44 Marburg Virus Disease incubation_period              <NA> Boris I..... 2014
#> 45 Marburg Virus Disease incubation_period              <NA> Boris I..... 2014
#> 46               Measles incubation_period             lnorm Justin L.... 2009
#> 47                  MERS incubation_period             lnorm Abdullah.... 2013
#> 48                  MERS incubation_period             gamma B J Cowl.... 2015
#> 49                  Mpox incubation_period              <NA> Fengying.... 2022
#> 50                  Mpox incubation_period              <NA> Fengying.... 2022
#> 51                  Mpox incubation_period              <NA> Fengying.... 2022
#> 52                  Mpox incubation_period             lnorm Fuminari.... 2022
#> 53                  Mpox incubation_period             gamma Giorgio .... 2022
#> 54                  Mpox incubation_period             lnorm Kelly Ch.... 2022
#> 55                  Mpox incubation_period              <NA> Shuqi Wa.... 2022
#> 56                  Mpox incubation_period              <NA> Shuqi Wa.... 2022
#> 57                  Mpox incubation_period              <NA> Shuqi Wa.... 2022
#> 58                  Mpox incubation_period             lnorm Zachary .... 2022
#> 59                  Mpox incubation_period             lnorm Zachary .... 2022
#> 60         Parainfluenza incubation_period             lnorm Justin L.... 2009
#> 61            Rhinovirus incubation_period             lnorm Justin L.... 2009
#> 62     Rift Valley Fever incubation_period             lnorm Kara E R.... 2014
#> 63                   RSV incubation_period             lnorm Justin L.... 2009
#> 64                   RSV incubation_period             lnorm Nicholas.... 2009
#> 65                   RSV incubation_period             lnorm Nicholas.... 2009
#> 66                  SARS incubation_period             lnorm Justin L.... 2009
#> 67       West Nile Fever incubation_period             lnorm Kara E R.... 2014
#> 68       West Nile Fever incubation_period             lnorm Kara E R.... 2014
#> 69       West Nile Fever incubation_period             lnorm Kara E R.... 2014
#> 70          Yellow Fever incubation_period             lnorm Kara E R.... 2014
#> 71          Yellow Fever incubation_period             lnorm Kara E R.... 2014
#> 72    Zika Virus Disease incubation_period             lnorm Justin L.... 2016
#>    sample_size
#> 1           14
#> 2           21
#> 3         6241
#> 4        28675
#> 5         1357
#> 6         1269
#> 7           59
#> 8           19
#> 9         1453
#> 10          52
#> 11         158
#> 12          52
#> 13         181
#> 14          99
#> 15         108
#> 16          73
#> 17         169
#> 18         124
#> 19         146
#> 20         146
#> 21         153
#> 22         196
#> 23        1798
#> 24          49
#> 25         957
#> 26         792
#> 27          13
#> 28         316
#> 29          16
#> 30          72
#> 31          72
#> 32         124
#> 33         151
#> 34          90
#> 35          78
#> 36         151
#> 37         151
#> 38         395
#> 39         173
#> 40         222
#> 41         229
#> 42         229
#> 43           6
#> 44          76
#> 45          18
#> 46          55
#> 47          23
#> 48         166
#> 49          NA
#> 50          NA
#> 51          NA
#> 52          18
#> 53         255
#> 54          22
#> 55          16
#> 56          27
#> 57         114
#> 58          35
#> 59          36
#> 60          11
#> 61          28
#> 62          23
#> 63          24
#> 64          24
#> 65          24
#> 66         157
#> 67          18
#> 68           8
#> 69           6
#> 70          91
#> 71          80
#> 72          25
# this same process can be achieved when loading the library
eparam <- epiparam(epi_dist = "incubation_period")

# filtering for onset to death
list_distributions(epiparam = eparam, epi_dist = "onset_to_death")
#> [1] disease           epi_distribution  prob_distribution author           
#> [5] year              sample_size      
#> <0 rows> (or 0-length row.names)