List distributions

List distributions included in package

Usage

list_distributions(
  delay_dist = c("all", "incubation", "onset_to_admission", "onset_to_death",
    "serial_interval", "generation_time"),
  parameters = FALSE
)

Arguments

delay_dist

A character defining parameter to be listed: "all" "incubation", "onset_to_admission", "onset_to_death", or "serial_interval". "all" will not filter by a certain delay distribution and return all records. "all" is the default.

parameters

A logical defining whether to show parameter values, default is FALSE

Author

Adam Kucharski

Examples

# the default for list_distributions() without any arguments is to return the
# incubation period
list_distributions()
#>             pathogen_id            type_id            study_id year size
#> 1            adenovirus    generation_time            Guo_etal 2020  375
#> 2            adenovirus         incubation        Lessler_etal 2009   14
#> 3                 ebola         incubation            WHO_team 2014  500
#> 4                 ebola     onset_to_death Ebola_outbreak_team 2018   14
#> 5                 ebola    serial_interval   WHO_response_team 2014   92
#> 6             human_CoV         incubation        Lessler_etal 2009   13
#> 7  influenza_A_seasonal         incubation        Lessler_etal 2009  122
#> 8  influenza_A_seasonal    Serial_interval           Levy_etal 2013  111
#> 9  influenza_B_seasonal         incubation        Lessler_etal 2009   76
#> 10 influenza_B_seasonal    Serial_interval           Levy_etal 2013   44
#> 11      influenza_H1N1p         incubation          Ghani_etal 2009   16
#> 12      influenza_H1N1p         incubation          Tuite_etal 2010  316
#> 13      influenza_H1N1p onset_to_admission           Jain_etal 2009  272
#> 14      influenza_H1N1p    Serial_interval           Levy_etal 2013  125
#> 15       influenza_H5N1         incubation        Cowling_etal 2013   27
#> 16       influenza_H5N1 onset_to_admission        Cowling_etal 2013   43
#> 17       influenza_H7N9         incubation        Cowling_etal 2013   32
#> 18       influenza_H7N9 onset_to_admission        Cowling_etal 2013  123
#> 19              marburg         incubation              Pavlin 2014   18
#> 20              measles         incubation        Lessler_etal 2009   56
#> 21             MERS_CoV         incubation         Assiri_etal 2013   23
#> 22             MERS_CoV         incubation     Cauchemez_et_al 2014    7
#> 23             MERS_CoV onset_to_admission         Assiri_etal 2013   23
#> 24             MERS_CoV    serial_interval         Assiri_etal 2013   23
#> 25            monkeypox         incubation               Nolen 2016   16
#> 26            monkeypox         incubation      Thornhill_etal 2022   23
#> 27            monkeypox    serial_interval               UKHSA 2022   17
#> 28        parainfluenza         incubation        Lessler_etal 2009   11
#> 29           rhinovirus         incubation        Lessler_etal 2009   28
#> 30                  RSV         incubation        Lessler_etal 2009   24
#> 31                  RSV    Serial_interval           Vink_etal 2014   59
#> 32             SARS_CoV         incubation       Donnelly_etal 2003   57
#> 33             SARS_CoV         incubation        Lessler_etal 2009  157
#> 34             SARS_CoV onset_to_admission       Donnelly_etal 2003   57
#> 35  SARS_CoV_2_wildtype    generation_time       Ferretti_etal 2020  191
#> 36  SARS_CoV_2_wildtype         incubation        McAloon_etal 2020 1269
#> 37  SARS_CoV_2_wildtype onset_to_admission         Linton_etal 2020  155
#> 38  SARS_CoV_2_wildtype     onset_to_death         Linton_etal 2020   39
#> 39  SARS_CoV_2_wildtype    serial_interval       Nishiura_etal 2020   28
#> 40  SARS_CoV_2_wildtype    serial_interval       Nishiura_etal 2020   18
#>    distribution
#> 1       weibull
#> 2         lnorm
#> 3         gamma
#> 4         gamma
#> 5         gamma
#> 6         lnorm
#> 7         lnorm
#> 8         lnorm
#> 9         lnorm
#> 10        lnorm
#> 11        gamma
#> 12        lnorm
#> 13        lnorm
#> 14        lnorm
#> 15      weibull
#> 16      weibull
#> 17      weibull
#> 18        gamma
#> 19        gamma
#> 20        lnorm
#> 21        lnorm
#> 22        lnorm
#> 23        lnorm
#> 24        lnorm
#> 25        lnorm
#> 26        lnorm
#> 27        gamma
#> 28        lnorm
#> 29        lnorm
#> 30        lnorm
#> 31        lnorm
#> 32        gamma
#> 33        lnorm
#> 34        gamma
#> 35      weibull
#> 36        lnorm
#> 37        lnorm
#> 38        lnorm
#> 39        lnorm
#> 40      weibull

# metrics can be chosen using the `delay_dist` argument, for example
# incubation period or onset to death
list_distributions(delay_dist = "incubation")
#>             pathogen_id    type_id        study_id year size distribution
#> 1            adenovirus incubation    Lessler_etal 2009   14        lnorm
#> 2                 ebola incubation        WHO_team 2014  500        gamma
#> 3             human_CoV incubation    Lessler_etal 2009   13        lnorm
#> 4  influenza_A_seasonal incubation    Lessler_etal 2009  122        lnorm
#> 5  influenza_B_seasonal incubation    Lessler_etal 2009   76        lnorm
#> 6       influenza_H1N1p incubation      Ghani_etal 2009   16        gamma
#> 7       influenza_H1N1p incubation      Tuite_etal 2010  316        lnorm
#> 8        influenza_H5N1 incubation    Cowling_etal 2013   27      weibull
#> 9        influenza_H7N9 incubation    Cowling_etal 2013   32      weibull
#> 10              marburg incubation          Pavlin 2014   18        gamma
#> 11              measles incubation    Lessler_etal 2009   56        lnorm
#> 12             MERS_CoV incubation     Assiri_etal 2013   23        lnorm
#> 13             MERS_CoV incubation Cauchemez_et_al 2014    7        lnorm
#> 14            monkeypox incubation           Nolen 2016   16        lnorm
#> 15            monkeypox incubation  Thornhill_etal 2022   23        lnorm
#> 16        parainfluenza incubation    Lessler_etal 2009   11        lnorm
#> 17           rhinovirus incubation    Lessler_etal 2009   28        lnorm
#> 18                  RSV incubation    Lessler_etal 2009   24        lnorm
#> 19             SARS_CoV incubation   Donnelly_etal 2003   57        gamma
#> 20             SARS_CoV incubation    Lessler_etal 2009  157        lnorm
#> 21  SARS_CoV_2_wildtype incubation    McAloon_etal 2020 1269        lnorm
list_distributions(delay_dist = "onset_to_death")
#>           pathogen_id        type_id            study_id year size distribution
#> 1               ebola onset_to_death Ebola_outbreak_team 2018   14        gamma
#> 2 SARS_CoV_2_wildtype onset_to_death         Linton_etal 2020   39        lnorm

# by default list_distributions() will return a filtered table, but if the
# full table including extra information (e.g. parameters of the
# distribution) the `parameters` argument can be set to `TRUE`
list_distributions(delay_dist = "onset_to_admission", parameters = TRUE)
#>           pathogen_id            type_id      study_id year size distribution
#> 1     influenza_H1N1p onset_to_admission     Jain_etal 2009  272        lnorm
#> 2      influenza_H5N1 onset_to_admission  Cowling_etal 2013   43      weibull
#> 3      influenza_H7N9 onset_to_admission  Cowling_etal 2013  123        gamma
#> 4            MERS_CoV onset_to_admission   Assiri_etal 2013   23        lnorm
#> 5            SARS_CoV onset_to_admission Donnelly_etal 2003   57        gamma
#> 6 SARS_CoV_2_wildtype onset_to_admission   Linton_etal 2020  155        lnorm
#>   mean    sd quantile_025 median quantile_75 quantile_875 quantile_95
#> 1   NA    NA           NA    3.0          NA           NA          NA
#> 2   NA    NA    1.3678290    4.9    6.770349           NA   10.222554
#> 3   NA    NA    0.6172617    4.2    6.170223           NA    9.622302
#> 4   NA    NA           NA    5.0          NA           NA          NA
#> 5 4.85  3.49           NA     NA          NA           NA          NA
#> 6 9.70 35.20           NA    2.6          NA           NA   35.100000
#>   quantile_975 lower_range upper_range    shape    scale   meanlog     sdlog
#> 1           NA          NA          NA       NA       NA 1.0986120 2.9115800
#> 2     11.42344          NA          NA 2.373396 6.438843        NA        NA
#> 3     10.82306          NA          NA 2.138094 2.153314        NA        NA
#> 4           NA           1          10       NA       NA 1.6089493 0.6103889
#> 5           NA          NA          NA 1.929655 2.513402        NA        NA
#> 6           NA          NA          NA       NA       NA 0.9466094 1.6281993
#>   extracted discretised phase_bias_adjusted
#> 1       yes        <NA>                <NA>
#> 2       yes          no                  no
#> 3       yes          no                  no
#> 4       yes        <NA>                <NA>
#> 5       yes        <NA>                <NA>
#> 6       yes          no                 yes
#>                                              notes     PMID
#> 1 Extracted using range 0 to 18, lognormal assumed 19815859
#> 2     extraction used quantile_025 and quantile_95 23803488
#> 3     extraction used quantile_025 and quantile_75 23803488
#> 4                                             <NA> 23782161
#> 5                                 Feb-26 to Mar-25 12781533
#> 6                              For living patients 32079150
#>                             DOI       added_by
#> 1         10.1056/NEJMoa0906695 Adam Kucharski
#> 2 10.1016/S0140-6736(13)61171-X Adam Kucharski
#> 3 10.1016/S0140-6736(13)61171-X Adam Kucharski
#> 4         10.1056/NEJMoa1306742 Adam Kucharski
#> 5 10.1016/S0140-6736(03)13410-1 Adam Kucharski
#> 6            10.3390/jcm9020538 Adam Kucharski