Print method for <multi_epidist> class

Print method for <multi_epidist> class

Usage

# S3 method for class 'multi_epidist'
print(x, ..., n = NULL)

Arguments

x

A <multi_epidist> object.

...

dots Extra arguments to be passed to the method.

n

A numeric specifying how many <epidist> objects to print. This argument is passed to head() for list printing. Default is NULL and the number of elements to print is controlled by package options().

Value

Invisibly returns a <multi_epidist>. Called for side-effects.

Examples

# entire database
db <- epidist_db()
#> Returning 122 results that match the criteria (99 are parameterised). 
#> Use subset to filter by entry variables or single_epidist to return a single entry. 
#> To retrieve the citation for each use the 'get_citation' function
db
#> # List of 122 <epidist> objects
#> Number of diseases: 23
#> ❯ Adenovirus ❯ COVID-19 ❯ Chikungunya ❯ Dengue ❯ Ebola Virus Disease ❯ Hantavirus Pulmonary Syndrome ❯ Human Coronavirus ❯ Influenza ❯ Japanese Encephalitis ❯ MERS ❯ Marburg Virus Disease ❯ Measles ❯ Mpox ❯ Parainfluenza ❯ Pneumonic Plague ❯ RSV ❯ Rhinovirus ❯ Rift Valley Fever ❯ SARS ❯ Smallpox ❯ West Nile Fever ❯ Yellow Fever ❯ Zika Virus Disease
#> Number of epi distributions: 12
#> ❯ generation time ❯ hospitalisation to death ❯ hospitalisation to discharge ❯ incubation period ❯ notification to death ❯ notification to discharge ❯ offspring distribution ❯ onset to death ❯ onset to discharge ❯ onset to hospitalisation ❯ onset to ventilation ❯ serial interval
#> [[1]]
#> Disease: Adenovirus
#> Pathogen: Adenovirus
#> Epi Distribution: incubation period
#> Study: Lessler J, Reich N, Brookmeyer R, Perl T, Nelson K, Cummings D (2009).
#> “Incubation periods of acute respiratory viral infections: a systematic
#> review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-6
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-6>.
#> Distribution: lnorm
#> Parameters:
#>   meanlog: 1.247
#>   sdlog: 0.975
#> 
#> [[2]]
#> Disease: Human Coronavirus
#> Pathogen: Human_Cov
#> Epi Distribution: incubation period
#> Study: Lessler J, Reich N, Brookmeyer R, Perl T, Nelson K, Cummings D (2009).
#> “Incubation periods of acute respiratory viral infections: a systematic
#> review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-7
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-7>.
#> Distribution: lnorm
#> Parameters:
#>   meanlog: 0.742
#>   sdlog: 0.918
#> 
#> [[3]]
#> Disease: SARS
#> Pathogen: SARS-Cov-1
#> Epi Distribution: incubation period
#> Study: Lessler J, Reich N, Brookmeyer R, Perl T, Nelson K, Cummings D (2009).
#> “Incubation periods of acute respiratory viral infections: a systematic
#> review.” _The Lancet Infectious Diseases_.
#> doi:10.1016/S1473-3099(09)70069-8
#> <https://doi.org/10.1016/S1473-3099%2809%2970069-8>.
#> Distribution: lnorm
#> Parameters:
#>   meanlog: 0.660
#>   sdlog: 1.205
#> 
#> # ℹ 119 more elements
#> # ℹ Use `print(n = ...)` to see more elements.
#> # ℹ Use `parameter_tbl()` to see a summary table of the parameters.
#> # ℹ Explore database online at: https://epiverse-trace.github.io/epiparameter/articles/database.html

# a single disease
db <- epidist_db(disease = "Ebola")
#> Returning 17 results that match the criteria (17 are parameterised). 
#> Use subset to filter by entry variables or single_epidist to return a single entry. 
#> To retrieve the citation for each use the 'get_citation' function
db
#> # List of 17 <epidist> objects
#> Number of diseases: 1
#> ❯ Ebola Virus Disease
#> Number of epi distributions: 9
#> ❯ hospitalisation to death ❯ hospitalisation to discharge ❯ incubation period ❯ notification to death ❯ notification to discharge ❯ offspring distribution ❯ onset to death ❯ onset to discharge ❯ serial interval
#> [[1]]
#> Disease: Ebola Virus Disease
#> Pathogen: Ebola Virus
#> Epi Distribution: offspring distribution
#> Study: Lloyd-Smith J, Schreiber S, Kopp P, Getz W (2005). “Superspreading and
#> the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> Distribution: nbinom
#> Parameters:
#>   mean: 1.500
#>   dispersion: 5.100
#> 
#> [[2]]
#> Disease: Ebola Virus Disease
#> Pathogen: Ebola Virus-Zaire Subtype
#> Epi Distribution: incubation period
#> Study: Eichner M, Dowell S, Firese N (2011). “Incubation period of ebola
#> hemorrhagic virus subtype zaire.” _Osong Public Health and Research
#> Perspectives_. doi:10.1016/j.phrp.2011.04.001
#> <https://doi.org/10.1016/j.phrp.2011.04.001>.
#> Distribution: lnorm
#> Parameters:
#>   meanlog: 2.487
#>   sdlog: 0.330
#> 
#> [[3]]
#> Disease: Ebola Virus Disease
#> Pathogen: Ebola Virus-Zaire Subtype
#> Epi Distribution: onset to death
#> Study: The Ebola Outbreak Epidemiology Team, Barry A, Ahuka-Mundeke S, Ali
#> Ahmed Y, Allarangar Y, Anoko J, Archer B, Abedi A, Bagaria J, Belizaire
#> M, Bhatia S, Bokenge T, Bruni E, Cori A, Dabire E, Diallo A, Diallo B,
#> Donnelly C, Dorigatti I, Dorji T, Waeber A, Fall I, Ferguson N,
#> FitzJohn R, Tengomo G, Formenty P, Forna A, Fortin A, Garske T,
#> Gaythorpe K, Gurry C, Hamblion E, Djingarey M, Haskew C, Hugonnet S,
#> Imai N, Impouma B, Kabongo G, Kalenga O, Kibangou E, Lee T, Lukoya C,
#> Ly O, Makiala-Mandanda S, Mamba A, Mbala-Kingebeni P, Mboussou F,
#> Mlanda T, Makuma V, Morgan O, Mulumba A, Kakoni P, Mukadi-Bamuleka D,
#> Muyembe J, Bathé N, Ndumbi Ngamala P, Ngom R, Ngoy G, Nouvellet P, Nsio
#> J, Ousman K, Peron E, Polonsky J, Ryan M, Touré A, Towner R, Tshapenda
#> G, Van De Weerdt R, Van Kerkhove M, Wendland A, Yao N, Yoti Z, Yuma E,
#> Kalambayi Kabamba G, Mwati J, Mbuy G, Lubula L, Mutombo A, Mavila O,
#> Lay Y, Kitenge E (2018). “Outbreak of Ebola virus disease in the
#> Democratic Republic of the Congo, April–May, 2018: an epidemiological
#> study.” _The Lancet_. doi:10.1016/S0140-6736(18)31387-4
#> <https://doi.org/10.1016/S0140-6736%2818%2931387-4>.
#> Distribution: gamma
#> Parameters:
#>   shape: 2.400
#>   scale: 3.333
#> 
#> # ℹ 14 more elements
#> # ℹ Use `print(n = ...)` to see more elements.
#> # ℹ Use `parameter_tbl()` to see a summary table of the parameters.
#> # ℹ Explore database online at: https://epiverse-trace.github.io/epiparameter/articles/database.html

# a single epi parameter
db <- epidist_db(epi_dist = "offspring distribution")
#> Returning 10 results that match the criteria (10 are parameterised). 
#> Use subset to filter by entry variables or single_epidist to return a single entry. 
#> To retrieve the citation for each use the 'get_citation' function
db
#> # List of 10 <epidist> objects
#> Number of diseases: 6
#> ❯ Ebola Virus Disease ❯ Hantavirus Pulmonary Syndrome ❯ Mpox ❯ Pneumonic Plague ❯ SARS ❯ Smallpox
#> Number of epi distributions: 1
#> ❯ offspring distribution
#> [[1]]
#> Disease: SARS
#> Pathogen: SARS-Cov-1
#> Epi Distribution: offspring distribution
#> Study: Lloyd-Smith J, Schreiber S, Kopp P, Getz W (2005). “Superspreading and
#> the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> Distribution: nbinom
#> Parameters:
#>   mean: 1.630
#>   dispersion: 0.160
#> 
#> [[2]]
#> Disease: SARS
#> Pathogen: SARS-Cov-1
#> Epi Distribution: offspring distribution
#> Study: Lloyd-Smith J, Schreiber S, Kopp P, Getz W (2005). “Superspreading and
#> the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> Distribution: nbinom
#> Parameters:
#>   mean: 0.940
#>   dispersion: 0.170
#> 
#> [[3]]
#> Disease: Smallpox
#> Pathogen: Smallpox-Variola-Major
#> Epi Distribution: offspring distribution
#> Study: Lloyd-Smith J, Schreiber S, Kopp P, Getz W (2005). “Superspreading and
#> the effect of individual variation on disease emergence.” _Nature_.
#> doi:10.1038/nature04153 <https://doi.org/10.1038/nature04153>.
#> Distribution: nbinom
#> Parameters:
#>   mean: 3.190
#>   dispersion: 0.370
#> 
#> # ℹ 7 more elements
#> # ℹ Use `print(n = ...)` to see more elements.
#> # ℹ Use `parameter_tbl()` to see a summary table of the parameters.
#> # ℹ Explore database online at: https://epiverse-trace.github.io/epiparameter/articles/database.html